refactor(core): decompose storage service responsibilities
