+++ framework/core/architecture/LAYERS.md
@@ -42,9 +42,9 @@ remain private implementation details behind the public `kungfu` facade.
 | `yijinjing-schema` | `schema-values` | `core/schema` | 5 | `yijinjing` | `src/libyijinjing/include/kungfu/yijinjing/schema/core.h` |
 | `yijinjing-kernel` | `journal-kernel` | `core/yijinjing` | 42 | `yijinjing` | `src/libyijinjing/include/kungfu/yijinjing/journal/journal.h`<br>`src/libyijinjing/include/kungfu/yijinjing/storage.h` |
 | `libkungfu-contracts` | `ports-contracts` | `core/runtime-contracts` | 55 | `yijinjing`<br>`kungfu_contracts`<br>`kungfu` | `src/libkungfu/include/kungfu/runtime/common.h`<br>`src/libkungfu/include/kungfu/runtime/storage/service.h` |
-| `libkungfu-services` | `application-services` | `core/runtime-services` | 30 | `kungfu_services`<br>`kungfu_services_state_cache`<br>`kungfu` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/state_service.cpp`<br>`src/libkungfu/src/runtime/query/fact_query.cpp` |
-| `libkungfu-adapters` | `adapters` | `core/runtime-adapters` | 13 | `kungfu_adapters`<br>`kungfu`<br>`kungfu_native_storage_shared` | `src/libkungfu/src/view/schema.cpp`<br>`src/libkungfu/src/runtime/native_storage.cpp`<br>`src/libkungfu/src/runtime/util/rocks.cpp` |
-| `core-composition-bindings` | `composition-bindings` | `core/bindings` | 42 | `kungfu_composition`<br>`kungfu`<br>`kungfu_embedding`<br>`kungfu_wasm_host`<br>`kungfu_node`<br>`kungfu_electron`<br>`drone`<br>`kungfu_kfc`<br>`kungfu_node_host`<br>`pykungfu` | `src/bindings/node/binding/kungfu_node.cpp`<br>`src/bindings/python/binding/pykungfu.cpp`<br>`src/libkungfu/src/runtime/embedding.cpp` |
+| `libkungfu-services` | `application-services` | `core/runtime-services` | 33 | `kungfu_services`<br>`kungfu_services_state_cache`<br>`kungfu` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/state_service.cpp`<br>`src/libkungfu/src/runtime/query/fact_query.cpp` |
+| `libkungfu-adapters` | `adapters` | `core/runtime-adapters` | 14 | `kungfu_adapters`<br>`kungfu`<br>`kungfu_native_storage_shared` | `src/libkungfu/src/view/schema.cpp`<br>`src/libkungfu/src/runtime/native_storage.cpp`<br>`src/libkungfu/src/runtime/util/rocks.cpp` |
+| `core-composition-bindings` | `composition-bindings` | `core/bindings` | 43 | `kungfu_composition`<br>`kungfu`<br>`kungfu_embedding`<br>`kungfu_wasm_host`<br>`kungfu_node`<br>`kungfu_electron`<br>`drone`<br>`kungfu_kfc`<br>`kungfu_node_host`<br>`pykungfu` | `src/bindings/node/binding/kungfu_node.cpp`<br>`src/bindings/python/binding/pykungfu.cpp`<br>`src/libkungfu/src/runtime/embedding.cpp` |
 | `core-native-qualification` | `qualification` | `core/qualification` | 16 | `yijinjing_mmap_tests`<br>`yijinjing_content_hash_tests`<br>`yijinjing_mmap_qualification`<br>`kungfu_durability_contract_tests`<br>`kungfu_runtime_error_tests`<br>`kungfu_embedding_generic_codec_tests`<br>`kungfu_peer_continuity_tests`<br>`kungfu_state_service_contract_tests`<br>`kungfu_durable_ingest_tests`<br>`kungfu_durability_powercut_fixture`<br>`kungfu_durability_slo_fixture`<br>`kungfu_offhost_backup_fixture`<br>`kungfu_projection_bootstrap_tests`<br>`kungfu_crash_recovery_tests`<br>`kungfu_profile_lifecycle_tests` | `src/libkungfu/tests/durability_contract_tests.cpp`<br>`src/libyijinjing/tests/mmap_tests.cpp` |
 
 ## Internal target graph
@@ -56,10 +56,23 @@ from the same authority as this map.
 | Target | Kind | Component | Depends on | Sources |
 | --- | --- | --- | --- | ---: |
 | `kungfu_contracts` | `INTERFACE` | `libkungfu-contracts` | — | 0 |
-| `kungfu_services` | `OBJECT` | `libkungfu-services` | `kungfu_contracts` | 27 |
+| `kungfu_services` | `OBJECT` | `libkungfu-services` | `kungfu_contracts` | 29 |
 | `kungfu_services_state_cache` | `OBJECT` | `libkungfu-services` | `kungfu_contracts` | 3 |
-| `kungfu_adapters` | `OBJECT` | `libkungfu-adapters` | `kungfu_services`<br>`kungfu_services_state_cache` | 12 |
-| `kungfu_composition` | `OBJECT` | `core-composition-bindings` | `kungfu_adapters` | 2 |
+| `kungfu_adapters` | `OBJECT` | `libkungfu-adapters` | `kungfu_services`<br>`kungfu_services_state_cache` | 13 |
+| `kungfu_composition` | `OBJECT` | `core-composition-bindings` | `kungfu_adapters` | 3 |
+
+## Responsibility seams
+
+These checked source budgets keep storage responsibilities from collapsing
+back into the compatibility facade.
+
+| Responsibility | Source | Line budget |
+| --- | --- | ---: |
+| Public storage facade, shared application helpers and compatibility surface | `src/libkungfu/src/runtime/storage/service.cpp` | 5000 |
+| File and RocksDB content providers behind the storage port | `src/libkungfu/src/runtime/storage/provider.cpp` | 800 |
+| Status, fsck, projection rebuild, GC and compaction planning | `src/libkungfu/src/runtime/storage/maintenance_service.cpp` | 500 |
+| Manifest bundle import, export and sync verification | `src/libkungfu/src/runtime/storage/transfer_service.cpp` | 520 |
+| JSON edge and domain operation dispatch composition | `src/libkungfu/src/runtime/storage/domain_dispatch.cpp` | 1000 |
 
 ## Navigation
 
@@ -70,8 +83,12 @@ from the same authority as this map.
 | State service and projections | `libkungfu-services` → `src/libkungfu/src/runtime/state_service.cpp` |
 | Runtime fact and saved-query behavior | `libkungfu-services` → `src/libkungfu/src/runtime/query/fact_query.cpp` |
 | Trust assessment or live coordination | `libkungfu-services` → `src/libkungfu/src/runtime/trust/assessment_runtime.cpp` |
-| Storage service composition | `libkungfu-services` → `src/libkungfu/src/runtime/storage/service.cpp` |
-| RocksDB or SQLite integration | `libkungfu-adapters` → `src/libkungfu/src/runtime/util/rocks.cpp` |
+| Storage facade or shared application helpers | `libkungfu-services` → `src/libkungfu/src/runtime/storage/service.cpp` |
+| Storage status, fsck, rebuild, GC or compact planning | `libkungfu-services` → `src/libkungfu/src/runtime/storage/maintenance_service.cpp` |
+| Storage bundle import, export or sync verification | `libkungfu-services` → `src/libkungfu/src/runtime/storage/transfer_service.cpp` |
+| File or RocksDB storage provider integration | `libkungfu-adapters` → `src/libkungfu/src/runtime/storage/provider.cpp` |
+| Storage JSON edge or domain operation dispatch | `core-composition-bindings` → `src/libkungfu/src/runtime/storage/domain_dispatch.cpp` |
+| General RocksDB or SQLite integration | `libkungfu-adapters` → `src/libkungfu/src/runtime/util/rocks.cpp` |
 | FlatBuffers projection boundary | `libkungfu-adapters` → `src/libkungfu/src/view/schema.cpp` |
 | Transport, process or OS integration | `libkungfu-adapters` → `src/libkungfu/src/runtime/io/io.cpp` |
 | Python, Node, C embedding or WASM entry | `core-composition-bindings` → `src/libkungfu/src/runtime/embedding.cpp` |
@@ -88,4 +105,5 @@ when a current target loses its CMake evidence, when a resolved internal
 include is undeclared, when a declared dependency or resolved include reverses
 the layer contract, when an internal source has zero or multiple build targets,
 when a target edge reverses the layer contract, when a forbidden include enters
-a protected layer, or when the map or generated CMake projection drifts.
+a protected layer, when a checked responsibility token or source-size budget
+drifts, or when the map or generated CMake projection drifts.
