+++ framework/core/src/libkungfu/src/runtime/storage/service.cpp
@@ -3,17 +3,14 @@
 #include <kungfu/runtime/storage/json_edge.h>
 #include <kungfu/runtime/storage/service.h>
 
+#include "service_internal.h"
+
 #include <algorithm>
 #include <array>
-#include <atomic>
 #include <charconv>
-#include <cstdlib>
 #include <filesystem>
-#include <fstream>
-#include <functional>
 #include <map>
 #include <memory>
-#include <mutex>
 #include <optional>
 #include <random>
 #include <stdexcept>
@@ -41,8 +38,6 @@
 #include <kungfu/yijinjing/storage/source_registry.h>
 #include <kungfu/yijinjing/storage/sync_root.h>
 #include <kungfu/yijinjing/time.h>
-#include <rocksdb/db.h>
-#include <rocksdb/iterator.h>
 #include <sqlite3.h>
 
 namespace kungfu::runtime::storage_service_api {
@@ -51,11 +46,8 @@ namespace fs = std::filesystem;
 namespace yy_storage = kungfu::yijinjing::storage;
 namespace yy_enums = kungfu::yijinjing::enums;
 
-namespace {
+namespace detail {
 
-inline constexpr const char *PAYLOAD_STATE_PRESENT = "present";
-inline constexpr const char *PAYLOAD_STATE_REDACTED = "redacted";
-inline constexpr const char *PAYLOAD_STATE_ABSENT = "absent";
 inline constexpr const char *CONTENT_TYPE_JSON = "application/json";
 
 nlohmann::json content_result_json(const yy_storage::content_store_result &result);
@@ -176,24 +168,7 @@ const char *payload_state_text(yy_enums::PayloadState state) {
   }
   return "missing";
 }
-inline constexpr const char *PROJECTION_SOURCE_REGISTRY = "source-registry-sqlite";
-inline constexpr const char *PROJECTION_MANIFEST_CATALOG = "manifest-catalog-sqlite";
-inline constexpr const char *PROVIDER_FILE = "content-addressed-file";
-inline constexpr const char *PROVIDER_ROCKSDB = "rocksdb";
-inline constexpr const char *ENV_STORAGE_PROVIDER = "KUNGFU_STORAGE_PROVIDER";
-
-struct provider_selection {
-  std::string name = {};
-  std::string source = {};
-};
-
-struct stored_payload {
-  std::string digest = {};
-  std::string uri = {};
-  uint64_t bytes = 0;
-};
-
-std::string text_or(const nlohmann::json &object, const std::string &field, const std::string &fallback = {}) {
+std::string text_or(const nlohmann::json &object, const std::string &field, const std::string &fallback) {
   if (!object.is_object() || !object.contains(field)) {
     return fallback;
   }
@@ -291,7 +266,7 @@ nlohmann::json array_or_empty(const nlohmann::json &object, const std::string &f
 
 std::string canonical_json(const nlohmann::json &value) { return value.dump(-1, ' ', false); }
 
-uint64_t uint64_or(const nlohmann::json &object, const std::string &field, uint64_t fallback = 0) {
+uint64_t uint64_or(const nlohmann::json &object, const std::string &field, uint64_t fallback) {
   if (!object.is_object() || !object.contains(field)) {
     return fallback;
   }
@@ -315,7 +290,7 @@ uint64_t uint64_or(const nlohmann::json &object, const std::string &field, uint6
   return fallback;
 }
 
-int64_t int64_or(const nlohmann::json &object, const std::string &field, int64_t fallback = 0) {
+int64_t int64_or(const nlohmann::json &object, const std::string &field, int64_t fallback) {
   if (!object.is_object() || !object.contains(field)) {
     return fallback;
   }
@@ -531,627 +506,6 @@ yy_storage::accepted_range_options parse_accepted_range_options(const nlohmann::
   return parsed;
 }
 
-fs::path root_dir(const std::string &runtime_dir) { return fs::path(runtime_dir) / "storage"; }
-
-fs::path payload_root(const std::string &runtime_dir) { return root_dir(runtime_dir) / "payloads"; }
-
-fs::path rocksdb_root(const std::string &runtime_dir) { return root_dir(runtime_dir) / "rocksdb"; }
-
-fs::path projection_root(const std::string &runtime_dir) { return root_dir(runtime_dir) / "projections"; }
-
-fs::path payload_path(const std::string &runtime_dir, const std::string &digest) {
-  // ADR-0037: payload bodies are opaque content-addressed bytes. The file is
-  // named by the content hash alone, with no format-implying extension — the
-  // body format is orthogonal to the record schema, which commits to the body
-  // by hash, length, and payload state (content_type is record metadata).
-  return payload_root(runtime_dir) / digest.substr(0, std::min<size_t>(2, digest.size())) / digest;
-}
-
-fs::path absolute_normalized(fs::path path) { return fs::absolute(std::move(path)).lexically_normal(); }
-
-fs::path runtime_home_path(const storage_service_options &options) {
-  const auto explicit_runtime_home = text_or(options.operation_options, "runtime_home");
-  if (!explicit_runtime_home.empty()) {
-    return absolute_normalized(explicit_runtime_home);
-  }
-  const auto runtime = absolute_normalized(options.runtime_dir);
-  return runtime.filename() == "runtime" ? runtime.parent_path() : runtime;
-}
-
-std::string runtime_home_source(const storage_service_options &options) {
-  return text_or(options.operation_options, "runtime_home").empty() ? "inferred-from-runtime-dir" : "option";
-}
-
-std::string optional_absolute_path(const nlohmann::json &object, const std::string &field) {
-  const auto value = text_or(object, field);
-  return value.empty() ? std::string{} : absolute_normalized(value).string();
-}
-
-std::vector<fs::path> all_payload_paths(const std::string &runtime_dir);
-std::string payload_digest_from_path(const fs::path &path);
-
-void write_json_file(const fs::path &path, const nlohmann::json &data) {
-  fs::create_directories(path.parent_path());
-  std::ofstream output(path, std::ios::trunc);
-  if (!output) {
-    throw std::runtime_error("failed to write JSON file: " + path.string());
-  }
-  output << data.dump(2, ' ', false) << '\n';
-}
-
-std::string read_bytes(const fs::path &path) {
-  std::ifstream input(path, std::ios::binary);
-  if (!input) {
-    throw std::runtime_error("failed to read payload: " + path.string());
-  }
-  return std::string(std::istreambuf_iterator<char>(input), std::istreambuf_iterator<char>());
-}
-
-void write_bytes(const fs::path &path, const std::string &raw) {
-  fs::create_directories(path.parent_path());
-  std::ofstream output(path, std::ios::binary | std::ios::trunc);
-  if (!output) {
-    throw std::runtime_error("failed to write payload: " + path.string());
-  }
-  output.write(raw.data(), static_cast<std::streamsize>(raw.size()));
-}
-
-std::string normalized_provider_name(const std::string &provider) {
-  if (provider.empty() || provider == "file" || provider == PROVIDER_FILE) {
-    return PROVIDER_FILE;
-  }
-  if (provider == "rocks" || provider == PROVIDER_ROCKSDB) {
-    return PROVIDER_ROCKSDB;
-  }
-  throw std::invalid_argument("unsupported storage provider: " + provider);
-}
-
-provider_selection select_provider(std::string provider) {
-  if (provider.empty()) {
-    if (const char *env_provider = std::getenv(ENV_STORAGE_PROVIDER); env_provider != nullptr) {
-      provider = env_provider;
-      return {normalized_provider_name(provider), "env:" + std::string(ENV_STORAGE_PROVIDER)};
-    }
-    return {PROVIDER_FILE, "default"};
-  }
-  return {normalized_provider_name(provider), "option"};
-}
-
-std::string storage_uri(const std::string &provider, const std::string &runtime_dir, const std::string &key) {
-  if (provider == PROVIDER_ROCKSDB) {
-    return std::string("rocksdb://") + rocksdb_root(runtime_dir).string() + "#" + key;
-  }
-  return key;
-}
-
-class storage_provider {
-public:
-  virtual ~storage_provider() = default;
-
-  [[nodiscard]] virtual std::string name() const = 0;
-  [[nodiscard]] virtual storage_provider_layout_view layout() const = 0;
-  [[nodiscard]] virtual storage_provider_runtime_view runtime() const = 0;
-  [[nodiscard]] virtual bool payload_exists(const std::string &digest) const = 0;
-  [[nodiscard]] virtual std::string read_payload(const std::string &digest) const = 0;
-  virtual void write_payload(const std::string &digest, const std::string &raw) const = 0;
-  [[nodiscard]] virtual std::vector<stored_payload> all_payloads() const = 0;
-  // ADR-0040: the immutable content store this provider publishes payload
-  // bytes through. Injected into kernel fsck/inspect so payload-ref
-  // resolution reads the same backend that owns the bytes.
-  [[nodiscard]] virtual yy_storage::content_store &content_store() const = 0;
-};
-
-class file_storage_provider : public storage_provider {
-public:
-  explicit file_storage_provider(std::string runtime_dir) : runtime_dir_(std::move(runtime_dir)) {}
-
-  [[nodiscard]] std::string name() const override { return PROVIDER_FILE; }
-
-  [[nodiscard]] storage_provider_layout_view layout() const override {
-    return {{},
-            "journal/system/storage/manifest-catalog/live/*.journal",
-            "storage/manifests/<hash-prefix>/<sha256>",
-            "storage/payloads/<hash-prefix>/<sha256>"};
-  }
-
-  [[nodiscard]] storage_provider_runtime_view runtime() const override {
-    return {"stateless-filesystem", "process-cached", "per filesystem operation", false, true};
-  }
-
-  [[nodiscard]] bool payload_exists(const std::string &digest) const override {
-    return fs::exists(payload_path(runtime_dir_, digest));
-  }
-
-  [[nodiscard]] std::string read_payload(const std::string &digest) const override {
-    return read_bytes(payload_path(runtime_dir_, digest));
-  }
-
-  void write_payload(const std::string &digest, const std::string &raw) const override {
-    // ADR-0040: publish through the immutable content store (atomic
-    // tmp+rename, digest checked against the bytes) instead of a bare file
-    // write; the store's layout is byte-compatible with payload_path.
-    const auto result = content_store_.put_if_absent("payloads", raw, yy_storage::make_content_hash(digest));
-    if (!result.ok()) {
-      throw std::runtime_error("failed to publish payload " + digest + ": " +
-                               yy_storage::content_store_error_name(result.error) +
-                               (result.message.empty() ? "" : " (" + result.message + ")"));
-    }
-  }
-
-  [[nodiscard]] yy_storage::content_store &content_store() const override { return content_store_; }
-
-  [[nodiscard]] std::vector<stored_payload> all_payloads() const override {
-    std::vector<stored_payload> result;
-    for (const auto &path : all_payload_paths(runtime_dir_)) {
-      result.push_back({payload_digest_from_path(path), path.string(), fs::file_size(path)});
-    }
-    return result;
-  }
-
-private:
-  std::string runtime_dir_;
-  mutable yy_storage::file_content_store content_store_{root_dir(runtime_dir_).string()};
-};
-
-// ADR-0040: the RocksDB-backed content store lives in the runtime/provider
-// layer and implements the yijinjing contract over the provider's single
-// long-lived engine handle (decision 6). Keys are "<namespace>/<digest>",
-// bare lowercase hex. The store never owns the handle: the provider does,
-// and multi-process ownership of one database path is rejected by the
-// engine's own lock. Values are written through the WAL in one atomic key
-// write, so a torn object is never visible under a digest; identical-bytes
-// races on the same key are benign under content identity.
-class rocksdb_content_store : public yy_storage::content_store {
-public:
-  // Returns a shared handle so an in-flight operation keeps its engine alive
-  // across a concurrent readonly-to-readwrite upgrade in the provider.
-  using engine_opener = std::function<std::shared_ptr<rocksdb::DB>(bool write)>;
-
-  explicit rocksdb_content_store(engine_opener open) : open_(std::move(open)) {}
-
-  [[nodiscard]] yy_storage::content_store_capabilities capabilities() const override {
-    yy_storage::content_store_capabilities caps{};
-    caps.profile = "kungfu-rocksdb/v1";
-    caps.hash_algorithm = yy_storage::CONTENT_HASH_ALGORITHM_SHA256;
-    caps.max_object_size = 0;
-    caps.atomic_put_if_absent = true;
-    caps.verified_reads = true;
-    caps.durability = write_options_.sync ? "fsync-per-write" : "wal-os-buffered";
-    caps.visibility = "publish-then-visible";
-    caps.concurrency = "multi-writer-single-process";
-    return caps;
-  }
-
-  [[nodiscard]] yy_storage::content_store_result put_if_absent(const std::string &content_namespace, const void *data,
-                                                               size_t size,
-                                                               const yy_storage::content_hash &expected) override {
-    yy_storage::content_store_result result{};
-    if (!yy_storage::is_valid_content_namespace(content_namespace)) {
-      result.error = yy_storage::content_store_error::InvalidArgument;
-      result.message = "invalid content namespace: " + content_namespace;
-      return result;
-    }
-    if (size > 0 && data == nullptr) {
-      result.error = yy_storage::content_store_error::InvalidArgument;
-      result.message = "null data with non-zero size";
-      return result;
-    }
-    const auto digest = yy_storage::compute_content_hash(data, size, yy_storage::CONTENT_HASH_ALGORITHM_SHA256);
-    if (!expected.empty()) {
-      result.error = yy_storage::validate_content_digest(expected, digest.algorithm, result.message);
-      if (result.error != yy_storage::content_store_error::Ok) {
-        return result;
-      }
-      if (expected.value != digest.value) {
-        result.error = yy_storage::content_store_error::HashMismatch;
-        result.message = "bytes hash to " + digest.value + ", caller declared " + expected.value;
-        return result;
-      }
-    }
-    result.hash = digest;
-    result.byte_length = size;
-    auto db = open_(true);
-    if (!db) {
-      result.error = yy_storage::content_store_error::IoError;
-      result.message = "cannot open storage engine for write";
-      return result;
-    }
-    const auto key = object_key(content_namespace, digest.value);
-    std::string existing;
-    auto status = db->Get(read_options_, key, &existing);
-    if (status.ok()) {
-      if (existing.size() != size) {
-        result.error = yy_storage::content_store_error::CorruptObject;
-        result.message = "existing object holds " + std::to_string(existing.size()) + " bytes, content is " +
-                         std::to_string(size) + " bytes; run verify";
-        return result;
-      }
-      result.existed = true;
-      return result;
-    }
-    if (!status.IsNotFound()) {
-      result.error = yy_storage::content_store_error::IoError;
-      result.message = "engine read failed: " + status.ToString();
-      return result;
-    }
-    status = db->Put(write_options_, key, rocksdb::Slice(static_cast<const char *>(data), size));
-    if (!status.ok()) {
-      result.error = yy_storage::content_store_error::IoError;
-      result.message = "engine write failed: " + status.ToString();
-      return result;
-    }
-    return result;
-  }
-
-  using yy_storage::content_store::put_if_absent;
-
-  [[nodiscard]] bool has(const std::string &content_namespace, const yy_storage::content_hash &hash) const override {
-    if (!yy_storage::is_valid_content_namespace(content_namespace)) {
-      return false;
-    }
-    std::string message;
-    if (yy_storage::validate_content_digest(hash, yy_storage::CONTENT_HASH_ALGORITHM_SHA256, message) !=
-        yy_storage::content_store_error::Ok) {
-      return false;
-    }
-    auto db = open_(false);
-    if (!db) {
-      return false;
-    }
-    std::string existing;
-    return db->Get(read_options_, object_key(content_namespace, hash.value), &existing).ok();
-  }
-
-  [[nodiscard]] yy_storage::content_store_result verify(const std::string &content_namespace,
-                                                        const yy_storage::content_hash &hash) const override {
-    yy_storage::content_store_result result{};
-    std::string bytes;
-    result.error = load_object(content_namespace, hash, bytes, result.message);
-    if (result.error != yy_storage::content_store_error::Ok) {
-      return result;
-    }
-    result.hash = yy_storage::make_content_hash(hash.value, yy_storage::CONTENT_HASH_ALGORITHM_SHA256);
-    result.byte_length = bytes.size();
-    if (!yy_storage::verify_content_hash(bytes, result.hash)) {
-      result.error = yy_storage::content_store_error::CorruptObject;
-      result.message = "stored bytes do not hash to " + result.hash.value;
-    }
-    return result;
-  }
-
-  [[nodiscard]] yy_storage::content_get_result get(const std::string &content_namespace,
-                                                   const yy_storage::content_hash &hash) const override {
-    yy_storage::content_get_result result{};
-    std::string bytes;
-    result.error = load_object(content_namespace, hash, bytes, result.message);
-    if (result.error != yy_storage::content_store_error::Ok) {
-      return result;
-    }
-    result.hash = yy_storage::make_content_hash(hash.value, yy_storage::CONTENT_HASH_ALGORITHM_SHA256);
-    if (!yy_storage::verify_content_hash(bytes, result.hash)) {
-      result.error = yy_storage::content_store_error::CorruptObject;
-      result.message = "stored bytes do not hash to " + result.hash.value;
-      return result;
-    }
-    result.bytes = std::move(bytes);
-    return result;
-  }
-
-private:
-  [[nodiscard]] static std::string object_key(const std::string &content_namespace, const std::string &digest) {
-    return content_namespace + "/" + digest;
-  }
-
-  [[nodiscard]] yy_storage::content_store_error load_object(const std::string &content_namespace,
-                                                            const yy_storage::content_hash &hash, std::string &bytes,
-                                                            std::string &message) const {
-    if (!yy_storage::is_valid_content_namespace(content_namespace)) {
-      message = "invalid content namespace: " + content_namespace;
-      return yy_storage::content_store_error::InvalidArgument;
-    }
-    const auto digest_error =
-        yy_storage::validate_content_digest(hash, yy_storage::CONTENT_HASH_ALGORITHM_SHA256, message);
-    if (digest_error != yy_storage::content_store_error::Ok) {
-      return digest_error;
-    }
-    auto db = open_(false);
-    if (!db) {
-      message = "no storage engine at this runtime dir";
-      return yy_storage::content_store_error::NotFound;
-    }
-    const auto status = db->Get(read_options_, object_key(content_namespace, hash.value), &bytes);
-    if (status.IsNotFound()) {
-      message = "no object under " + object_key(content_namespace, hash.value);
-      return yy_storage::content_store_error::NotFound;
-    }
-    if (!status.ok()) {
-      message = "engine read failed: " + status.ToString();
-      return yy_storage::content_store_error::IoError;
-    }
-    return yy_storage::content_store_error::Ok;
-  }
-
-  engine_opener open_;
-  rocksdb::ReadOptions read_options_ = [] {
-    rocksdb::ReadOptions options;
-    options.fill_cache = false;
-    return options;
-  }();
-  rocksdb::WriteOptions write_options_ = {};
-};
-
-class rocksdb_storage_provider : public storage_provider {
-public:
-  explicit rocksdb_storage_provider(std::string runtime_dir)
-      : runtime_dir_(std::move(runtime_dir)), content_store_([this](bool write) { return open(write); }) {}
-
-  [[nodiscard]] std::string name() const override { return PROVIDER_ROCKSDB; }
-
-  [[nodiscard]] storage_provider_layout_view layout() const override {
-    return {"storage/rocksdb", "journal/system/storage/manifest-catalog/live/*.journal", "manifests/<sha256>",
-            "payloads/<sha256>"};
-  }
-
-  [[nodiscard]] storage_provider_runtime_view runtime() const override {
-    std::lock_guard<std::mutex> lock(db_mutex_);
-    return {"provider-instance-owned",
-            "process-cached",
-            db_ ? (db_writable_ ? "open-readwrite" : "open-readonly") : "closed",
-            false,
-            true,
-            read_options_.fill_cache,
-            write_options_.sync};
-  }
-
-  [[nodiscard]] bool payload_exists(const std::string &digest) const override {
-    return content_store_.has("payloads", yy_storage::make_content_hash(digest));
-  }
-
-  [[nodiscard]] std::string read_payload(const std::string &digest) const override {
-    // verified read through the content store: corrupt bytes never come back
-    auto result = content_store_.get("payloads", yy_storage::make_content_hash(digest));
-    if (!result.ok()) {
-      throw std::runtime_error("failed to read payload " + digest + ": " +
-                               yy_storage::content_store_error_name(result.error) +
-                               (result.message.empty() ? "" : " (" + result.message + ")"));
-    }
-    return std::move(result.bytes);
-  }
-
-  void write_payload(const std::string &digest, const std::string &raw) const override {
-    const auto result = content_store_.put_if_absent("payloads", raw, yy_storage::make_content_hash(digest));
-    if (!result.ok()) {
-      throw std::runtime_error("failed to publish payload " + digest + ": " +
-                               yy_storage::content_store_error_name(result.error) +
-                               (result.message.empty() ? "" : " (" + result.message + ")"));
-    }
-  }
-
-  [[nodiscard]] std::vector<stored_payload> all_payloads() const override {
-    std::vector<stored_payload> result;
-    for_each("payloads/", [&](const std::string &key, const std::string &raw) {
-      result.push_back({key.substr(std::string("payloads/").size()), uri_for(key), raw.size()});
-    });
-    std::sort(result.begin(), result.end(),
-              [](const stored_payload &lhs, const stored_payload &rhs) { return lhs.digest < rhs.digest; });
-    return result;
-  }
-
-  [[nodiscard]] yy_storage::content_store &content_store() const override { return content_store_; }
-
-private:
-  [[nodiscard]] std::string uri_for(const std::string &key) const {
-    return storage_uri(PROVIDER_ROCKSDB, runtime_dir_, key);
-  }
-
-  // Lazily opens the engine and hands back a shared handle. RocksDB is
-  // thread-safe through one handle, so concurrent operations share it; the
-  // readonly-to-readwrite upgrade swaps in a fresh handle while in-flight
-  // readers finish on the old one (a readonly open holds no engine lock).
-  // Readonly opens still never create the database; only writes do.
-  [[nodiscard]] std::shared_ptr<rocksdb::DB> open(bool write) const {
-    std::lock_guard<std::mutex> lock(db_mutex_);
-    if (db_) {
-      if (!write || db_writable_) {
-        return db_;
-      }
-      db_.reset();
-      db_writable_ = false;
-    }
-    rocksdb::DB *raw = nullptr;
-    rocksdb::Options options;
-    options.create_if_missing = write;
-    options.error_if_exists = false;
-    rocksdb::Status status;
-    if (write) {
-      fs::create_directories(rocksdb_root(runtime_dir_));
-      status = rocksdb::DB::Open(options, rocksdb_root(runtime_dir_).string(), &raw);
-    } else {
-      if (!fs::exists(rocksdb_root(runtime_dir_))) {
-        return {};
-      }
-      status = rocksdb::DB::OpenForReadOnly(options, rocksdb_root(runtime_dir_).string(), &raw);
-    }
-    if (!status.ok()) {
-      throw std::runtime_error("rocksdb_open_failed: " + status.ToString());
-    }
-    db_ = std::shared_ptr<rocksdb::DB>(raw);
-    db_writable_ = write;
-    return db_;
-  }
-
-  [[nodiscard]] bool get(const std::string &key, std::string &value) const {
-    auto db = open(false);
-    if (!db) {
-      return false;
-    }
-    const auto status = db->Get(read_options_, key, &value);
-    if (status.IsNotFound()) {
-      return false;
-    }
-    if (!status.ok()) {
-      throw std::runtime_error("rocksdb_read_failed: " + key + ": " + status.ToString());
-    }
-    return true;
-  }
-
-  void put(const std::string &key, const std::string &value) const {
-    auto db = open(true);
-    const auto status = db->Put(write_options_, key, value);
-    if (!status.ok()) {
-      throw std::runtime_error("rocksdb_write_failed: " + key + ": " + status.ToString());
-    }
-  }
-
-  template <typename Fn> void for_each(const std::string &prefix, Fn fn) const {
-    auto db = open(false);
-    if (!db) {
-      return;
-    }
-    std::unique_ptr<rocksdb::Iterator> it(db->NewIterator(read_options_));
-    for (it->Seek(prefix); it->Valid(); it->Next()) {
-      const auto key = it->key().ToString();
-      if (!key.starts_with(prefix)) {
-        break;
-      }
-      fn(key, it->value().ToString());
-    }
-    if (!it->status().ok()) {
-      throw std::runtime_error("rocksdb_iterate_failed: " + it->status().ToString());
-    }
-  }
-
-  std::string runtime_dir_;
-  mutable rocksdb_content_store content_store_;
-  mutable std::mutex db_mutex_;
-  mutable std::shared_ptr<rocksdb::DB> db_ = {};
-  mutable bool db_writable_ = false;
-  rocksdb::ReadOptions read_options_ = [] {
-    rocksdb::ReadOptions options;
-    options.fill_cache = false;
-    return options;
-  }();
-  rocksdb::WriteOptions write_options_ = {};
-};
-
-std::unique_ptr<storage_provider> make_provider(const std::string &provider_name, const std::string &runtime_dir) {
-  if (provider_name == PROVIDER_ROCKSDB) {
-    return std::make_unique<rocksdb_storage_provider>(runtime_dir);
-  }
-  return std::make_unique<file_storage_provider>(runtime_dir);
-}
-
-// ADR-0040 decision 6: the per-operation provider open/close was a lifecycle
-// artifact, not an engine limit. One long-lived provider per (canonical
-// runtime dir, provider) is shared by every operation in this process, so
-// concurrent facade/service calls share one engine handle instead of racing
-// for the engine lock. Entries live until process exit: the touched
-// (runtime dir, provider) set is small, an evicted-then-reused handle would
-// reintroduce the open/close races this cache removes, and a background
-// eviction thread is out of scope by design. The engine's own lock keeps
-// rejecting a second process on the same database path — holding the write
-// handle for the process lifetime is that decision made visible, and reactor's
-// location-metadata engine lives under layout::MAP, a disjoint path from
-// this provider's storage/rocksdb, so no path ever has two in-process owners.
-class provider_cache {
-public:
-  static provider_cache &instance() {
-    // Intentionally leaked: cached engine handles must not run destructors
-    // during static teardown (RocksDB aborts once its lock infrastructure is
-    // torn down first). Never closing on exit loses nothing under the
-    // declared contract — publication is WAL-ordered, so exit-without-close
-    // is exactly the crash-safety case the backend already commits to.
-    static auto *cache = new provider_cache();
-    return *cache;
-  }
-
-  [[nodiscard]] std::shared_ptr<storage_provider> acquire(const std::string &runtime, const std::string &provider) {
-    const auto selection = select_provider(provider);
-    const auto runtime_dir = absolute_normalized(runtime).string();
-    const auto key = selection.name + "|" + runtime_dir;
-    std::lock_guard<std::mutex> lock(mutex_);
-    if (const auto it = providers_.find(key); it != providers_.end()) {
-      hits_.fetch_add(1, std::memory_order_relaxed);
-      return it->second;
-    }
-    misses_.fetch_add(1, std::memory_order_relaxed);
-    return providers_.emplace(key, make_provider(selection.name, runtime_dir)).first->second;
-  }
-
-  [[nodiscard]] storage_provider_cache_view stats() const {
-    std::lock_guard<std::mutex> lock(mutex_);
-    return {"process", providers_.size(), hits_.load(std::memory_order_relaxed),
-            misses_.load(std::memory_order_relaxed)};
-  }
-
-private:
-  provider_cache() = default;
-
-  mutable std::mutex mutex_;
-  std::unordered_map<std::string, std::shared_ptr<storage_provider>> providers_;
-  std::atomic<uint64_t> hits_{0};
-  std::atomic<uint64_t> misses_{0};
-};
-
-nlohmann::json provider_runtime_json(const storage_provider_runtime_view &runtime) {
-  nlohmann::json rendered = {{"lifecycle", runtime.lifecycle},
-                             {"instance_lifecycle", runtime.instance_lifecycle},
-                             {"handle", runtime.handle},
-                             {"readonly_open_creates_backend", runtime.readonly_open_creates_backend},
-                             {"write_open_creates_backend", runtime.write_open_creates_backend}};
-  if (runtime.read_fill_cache.has_value()) {
-    rendered["read_options"] = {{"fill_cache", *runtime.read_fill_cache}};
-  }
-  if (runtime.write_sync.has_value()) {
-    rendered["write_options"] = {{"sync", *runtime.write_sync}};
-  }
-  return rendered;
-}
-
-nlohmann::json provider_layout_json(const storage_provider_layout_view &layout) {
-  nlohmann::json rendered = {{"manifest_catalog_journal", layout.manifest_catalog_journal},
-                             {"manifest_entries", layout.manifest_entries},
-                             {"payloads", layout.payloads}};
-  if (layout.database.has_value()) {
-    rendered["database"] = *layout.database;
-  }
-  return rendered;
-}
-
-nlohmann::json provider_cache_json(const storage_provider_cache_view &cache) {
-  return {{"lifecycle", cache.lifecycle}, {"entries", cache.entries}, {"hits", cache.hits}, {"misses", cache.misses}};
-}
-
-std::shared_ptr<storage_provider> shared_provider(const storage_service_options &options) {
-  return provider_cache::instance().acquire(options.runtime_dir, options.provider);
-}
-
-std::shared_ptr<storage_provider> shared_provider(const std::string &runtime_dir) {
-  return provider_cache::instance().acquire(runtime_dir, {});
-}
-
-// Bundle a provider with an episode store wired to its content store, so
-// payload-ref resolution reads the same backend that published the bytes
-// (ADR-0040); the provider member keeps the injected store alive.
-struct episode_store_with_provider {
-  std::shared_ptr<storage_provider> provider;
-  yy_storage::episode_manifest_store store;
-};
-
-episode_store_with_provider episode_ref_store(const storage_service_options &options) {
-  auto provider = shared_provider(options);
-  auto store = yy_storage::episode_manifest_store(options.runtime_dir);
-  store.set_content_store(&provider->content_store());
-  return {std::move(provider), std::move(store)};
-}
-
-episode_store_with_provider episode_ref_store(const storage_fsck_request &request) {
-  auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-  auto store = yy_storage::episode_manifest_store(request.runtime_dir);
-  store.set_content_store(&provider->content_store());
-  return {std::move(provider), std::move(store)};
-}
-
 storage_layout_result workspace_episode_layout_typed(const storage_layout_request &request,
                                                      const storage_provider &provider) {
   const auto runtime = absolute_normalized(request.runtime_dir);
@@ -1189,7 +543,7 @@ storage_layout_result workspace_episode_layout_typed(const storage_layout_reques
                   (manifest_catalog_journal_dir / "*.journal").string(),
                   manifest_entries_pattern.string(),
                   payload_pattern.string(),
-                  rocksdb_root(runtime.string()).string(),
+                  provider_database_path(runtime.string()).string(),
                   source_registry_projection(runtime.string()).sqlite_path(),
                   manifest_catalog_projection(runtime.string()).sqlite_path(),
                   episode_manifest_dir.string(),
@@ -1290,34 +644,6 @@ nlohmann::json entries_for_manifest(const nlohmann::json &manifest, const nlohma
     }
   }
   return result;
-}
-
-std::vector<fs::path> all_payload_paths(const std::string &runtime_dir) {
-  std::vector<fs::path> paths;
-  const auto root = payload_root(runtime_dir);
-  if (!fs::exists(root)) {
-    return paths;
-  }
-  for (const auto &prefix : fs::directory_iterator(root)) {
-    if (!prefix.is_directory()) {
-      continue;
-    }
-    for (const auto &entry : fs::directory_iterator(prefix.path())) {
-      // Payload bodies are opaque content-addressed files named by hash, with no
-      // extension (ADR-0037); every regular file under a prefix is a body.
-      if (entry.is_regular_file()) {
-        paths.emplace_back(entry.path());
-      }
-    }
-  }
-  std::sort(paths.begin(), paths.end());
-  return paths;
-}
-
-std::string payload_digest_from_path(const fs::path &path) {
-  // Bodies are named by the full content hash with no extension; the whole
-  // filename is the digest.
-  return path.filename().string();
 }
 
 yy_storage::manifest_catalog_store catalog_store(const std::string &runtime_dir) {
@@ -3824,1533 +3150,6 @@ nlohmann::json source_registry_record_json(const yijinjing::types::AcceptedRange
           {"status", verification_status_text(record.status)}};
 }
 
-storage_status_result status_typed_impl(const storage_status_request &request) {
-  const auto selection = select_provider(request.provider);
-  const auto provider = provider_cache::instance().acquire(request.runtime_dir, selection.name);
-  storage_status_result result{};
-  result.backend = provider->name();
-  result.provider = provider->name();
-  result.provider_config_source =
-      request.provider_config_source.empty() ? selection.source : request.provider_config_source;
-  result.provider_runtime = provider->runtime();
-  result.scope = request.source_id.empty() ? "all" : "source";
-  if (!request.source_id.empty()) {
-    result.source_id = request.source_id;
-  }
-
-  const auto source_fold = registry_store(request.runtime_dir).fold_typed_records();
-  for (const auto &[source_uid, source] : source_fold.sources) {
-    (void)source_uid;
-    auto view = source_registry_status_view(source);
-    if (request.source_id.empty() || view.source_id == request.source_id) {
-      result.sources.push_back(std::move(view));
-    }
-  }
-  result.ok = request.source_id.empty() || !result.sources.empty();
-
-  const auto catalog = catalog_store(request.runtime_dir).read_typed_records();
-  std::unordered_map<uint64_t, const yijinjing::types::ImportManifestAccepted *> latest_manifests;
-  std::unordered_map<uint64_t, const yijinjing::types::ChannelCursorUpdated *> latest_cursors;
-  for (const auto &manifest : catalog.manifests) {
-    latest_manifests[manifest.source_uid] = &manifest;
-  }
-  for (const auto &cursor : catalog.cursors) {
-    latest_cursors[cursor.source_uid] = &cursor;
-  }
-
-  for (const auto &source : result.sources) {
-    storage_source_status_view status{};
-    status.source_id = source.source_id;
-    status.source = source;
-    const auto manifest_iter = latest_manifests.find(source.source_uid);
-    if (manifest_iter == latest_manifests.end()) {
-      status.reason = "manifest_missing";
-      result.source_status.push_back(std::move(status));
-      continue;
-    }
-
-    const auto &manifest = *manifest_iter->second;
-    status.ok = true;
-    status.manifest_id = fixed_string(manifest.manifest_id);
-    status.source_type = fixed_string(manifest.source_type);
-    status.source_head = fixed_string(manifest.source_head);
-    status.accepted_range = accepted_range_status_view(manifest);
-    status.sync_root =
-        storage_sync_root_view{fixed_string(manifest.sync_root_algo), fixed_string(manifest.sync_root_value)};
-    status.entries = manifest.entry_count;
-    status.payload_inventory = manifest.entry_count;
-
-    std::unordered_set<std::string> schema_keys;
-    for (const auto &entry : catalog.entries) {
-      if (entry.manifest_uid != manifest.manifest_uid || entry.accept_time != manifest.accept_time ||
-          entry.entry_schema_version == 0) {
-        continue;
-      }
-      schema_keys.insert(fixed_string(entry.kind) + ":" + std::to_string(entry.entry_schema_version));
-    }
-    status.schema_inventory = schema_keys.size();
-
-    if (const auto cursor_iter = latest_cursors.find(source.source_uid); cursor_iter != latest_cursors.end()) {
-      status.accepted_cursor = cursor_status_view(*cursor_iter->second);
-    }
-    const auto source_type = *status.source_type;
-    status.source_record =
-        storage_manifest_source_view{status.source_id,
-                                     source_type,
-                                     source_type == "atlas" ? "adapter" : "local",
-                                     fixed_string(manifest.source_coordinate),
-                                     *status.source_head,
-                                     {fixed_string(manifest.range_since), fixed_string(manifest.range_until)},
-                                     fixed_string(manifest.sync_root_value),
-                                     *status.accepted_range,
-                                     *status.manifest_id};
-    result.source_status.push_back(std::move(status));
-  }
-
-  result.projections = {source_registry_projection_status(request.runtime_dir),
-                        manifest_catalog_projection_status(request.runtime_dir)};
-  result.provider_cache = provider_cache::instance().stats();
-  return result;
-}
-
-std::vector<std::string> referenced_payload_hashes(const std::string &runtime_dir, const std::string &source_id = {}) {
-  return catalog_store(runtime_dir).referenced_payload_hashes(source_id);
-}
-
-std::pair<nlohmann::json, std::string> load_payload_impl(const storage_provider &provider,
-                                                         const nlohmann::json &entry) {
-  const auto digest = text_or(entry, "payload_hash");
-  if (digest.empty() || !provider.payload_exists(digest)) {
-    // An empty digest can never resolve to a body (and must not reach the
-    // hash primitives, which reject empty values).
-    return {nullptr, "payload_missing"};
-  }
-  const auto raw = provider.read_payload(digest);
-  if (!entry.contains("byte_len") ||
-      !entry.at("byte_len").is_number_unsigned() && !entry.at("byte_len").is_number_integer()) {
-    return {nullptr, "byte_len_mismatch"};
-  }
-  const auto expected_len = entry.at("byte_len").get<uint64_t>();
-  const auto error =
-      yy_storage::verify_payload_ref(raw, digest, expected_len, yy_storage::CONTENT_HASH_ALGORITHM_SHA256);
-  if (!error.empty()) {
-    return {nullptr, error};
-  }
-  try {
-    auto payload = nlohmann::json::parse(raw);
-    return {payload, {}};
-  } catch (const nlohmann::json::exception &) {
-    return {nullptr, "payload_decode_error"};
-  }
-}
-
-storage_fsck_result fsck_typed_impl(const storage_fsck_request &request) {
-  if (request.scope == storage_fsck_scope::Episode)
-    return episode_fsck_typed_impl(request);
-
-  storage_fsck_result result{};
-  result.scope = request.source_id.empty() ? storage_fsck_scope::All : storage_fsck_scope::Source;
-  if (!request.source_id.empty())
-    result.source_id = request.source_id;
-  result.checked.projection_indexes = 2;
-
-  const auto status =
-      status_typed_impl({request.runtime_dir, request.provider, request.provider_config_source, request.source_id});
-  std::vector<storage_source_registry_view> sources;
-  for (const auto &source : status.sources) {
-    if (source.registered)
-      sources.push_back(source);
-  }
-  result.checked.sources = sources.size();
-  if (!request.source_id.empty() && sources.empty()) {
-    storage_fsck_cross_issue detail{};
-    detail.source_id = request.source_id;
-    result.issues.push_back({"error", "source_missing", "source-registry", detail});
-    result.ok = false;
-    result.status = "failed";
-    return result;
-  }
-
-  result.source_registry = registry_store(request.runtime_dir).fsck_typed(request.source_id);
-  for (const auto &error : result.source_registry.errors) {
-    if (error.code != "source_missing")
-      result.issues.push_back({"error", error.code, "source-registry", error});
-  }
-  for (const auto &warning : result.source_registry.warnings)
-    result.issues.push_back({"warning", warning.code, "source-registry", warning});
-
-  const auto catalog_records = catalog_store(request.runtime_dir).read_typed_records();
-  std::unordered_map<uint64_t, const yijinjing::types::ImportManifestAccepted *> latest_manifests;
-  for (const auto &manifest : catalog_records.manifests)
-    latest_manifests[manifest.source_uid] = &manifest;
-  uint64_t sources_with_manifests = 0;
-  for (const auto &source : sources) {
-    const auto iter = latest_manifests.find(source.source_uid);
-    if (iter == latest_manifests.end()) {
-      storage_fsck_cross_issue detail{};
-      detail.source_id = source.source_id;
-      result.issues.push_back({"error", "manifest_missing", "manifest-catalog", detail});
-      continue;
-    }
-    ++sources_with_manifests;
-    const auto catalog_head = fixed_string(iter->second->source_head);
-    if (source.head.has_value() && !source.head->empty() && *source.head != catalog_head) {
-      storage_fsck_cross_issue detail{};
-      detail.source_id = source.source_id;
-      detail.expected = catalog_head;
-      detail.actual = *source.head;
-      result.issues.push_back({"error", "source_registry_drift", "source-registry", detail});
-    }
-    result.checked.accepted_ranges += source.accepted_range_count;
-  }
-  result.checked.source_records = sources_with_manifests;
-
-  auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-  if (sources_with_manifests > 0 || request.source_id.empty()) {
-    result.manifest_catalog =
-        catalog_store(request.runtime_dir).fsck_typed(request.source_id, provider->content_store());
-    for (const auto &error : result.manifest_catalog->errors) {
-      if (error.code != "source_missing")
-        result.issues.push_back({"error", error.code, "manifest-catalog", error});
-    }
-    for (const auto &warning : result.manifest_catalog->warnings)
-      result.issues.push_back({"warning", warning.code, "manifest-catalog", warning});
-    result.degraded = result.degraded || result.manifest_catalog->degraded;
-    result.checked.manifests = result.manifest_catalog->manifests;
-    result.checked.manifest_entries = result.manifest_catalog->manifest_entries;
-    result.checked.payloads = result.manifest_catalog->payloads;
-    result.checked.entries_documents = result.manifest_catalog->entries_documents;
-  }
-
-  if (request.source_id.empty()) {
-    const auto referenced = referenced_payload_hashes(request.runtime_dir);
-    for (const auto &payload : provider->all_payloads()) {
-      if (std::find(referenced.begin(), referenced.end(), payload.digest) == referenced.end()) {
-        ++result.checked.orphan_payloads;
-        storage_fsck_cross_issue detail{};
-        detail.path = payload.uri;
-        detail.payload_hash = payload.digest;
-        result.issues.push_back({"warning", "orphan_payload", "content-store", detail});
-      }
-    }
-  }
-
-  const auto scoped = episode_ref_store(request);
-  result.episode_manifest = scoped.store.fsck_typed();
-  result.checked.episode_manifest_records = result.episode_manifest.episode_manifest_records;
-  result.checked.episodes = result.episode_manifest.episodes;
-  for (const auto &error : result.episode_manifest.errors)
-    result.issues.push_back({"error", error.code, "episode-manifest", error});
-  for (const auto &warning : result.episode_manifest.warnings)
-    result.issues.push_back({"warning", warning.code, "episode-manifest", warning});
-  result.degraded = result.degraded || result.episode_manifest.degraded;
-
-  result.projections = {source_registry_projection_status(request.runtime_dir),
-                        manifest_catalog_projection_status(request.runtime_dir)};
-  for (const auto &projection : result.projections) {
-    if (projection.verification.status == "absent") {
-      result.issues.push_back({"warning", "projection_absent", projection.name, projection});
-    } else if (projection.verification.status == "degraded") {
-      result.degraded = true;
-      result.issues.push_back({"warning", "projection_drift", projection.name, projection});
-    }
-  }
-  result.ok = std::none_of(result.issues.begin(), result.issues.end(),
-                           [](const auto &issue) { return issue.severity == "error"; });
-  result.status = !result.ok ? "failed" : (result.degraded ? "degraded" : "ok");
-  return result;
-}
-
-nlohmann::json fsck_impl(const storage_service_options &options) {
-  return render_storage_fsck_result(default_storage_service().fsck(parse_storage_fsck_request(options)));
-}
-
-storage_rebuild_index_result rebuild_index_typed_impl(const storage_rebuild_index_request &request) {
-  // ADR-0037 (final slice): rebuild the derived SQLite projections from the
-  // kernel journals through the Hana closed-set -> SQLite path. The journals
-  // are the authority; there is no JSON registry to regenerate any more.
-  storage_rebuild_index_result result{};
-  result.scope = request.source_id.empty() ? "all" : "source";
-  if (!request.source_id.empty()) {
-    result.source_id = request.source_id;
-  }
-  result.dry_run = request.dry_run;
-  result.written = !request.dry_run;
-  const auto plan_one = [&](const char *name, auto &&projection) {
-    auto verify = projection.verify_typed();
-    const bool needs_write = verify.status != "ok" || !verify.projection_present;
-    result.would_write = result.would_write || needs_write;
-    if (request.dry_run) {
-      result.projections.push_back({name, true, false, needs_write, std::move(verify)});
-      return;
-    }
-    auto rebuilt = projection.rebuild_typed();
-    if (!rebuilt.ok) {
-      result.errors.push_back({"projection_rebuild_failed", std::string(name), {}});
-    }
-    result.projections.push_back({name, false, true, true, std::move(rebuilt)});
-  };
-  plan_one(PROJECTION_SOURCE_REGISTRY, source_registry_projection(request.runtime_dir));
-  plan_one(PROJECTION_MANIFEST_CATALOG, manifest_catalog_projection(request.runtime_dir));
-  const auto sources = registry_store(request.runtime_dir).fold_typed_records();
-  result.sources_rebuilt = sources.sources.size();
-  if (!request.source_id.empty()) {
-    const auto found = std::any_of(sources.sources.begin(), sources.sources.end(), [&](const auto &item) {
-      return item.second.registered && fixed_string(item.second.registration.source_id) == request.source_id;
-    });
-    if (!found) {
-      result.errors.push_back({"source_missing", {}, request.source_id});
-    }
-  }
-  result.ok = result.errors.empty();
-  if (!request.dry_run) {
-    result.would_write = true;
-  }
-  return result;
-}
-
-nlohmann::json rebuild_index_impl(const storage_service_options &options) {
-  return render_storage_rebuild_index_result(
-      default_storage_service().rebuild_index(parse_storage_rebuild_index_request(options)));
-}
-
-storage_gc_plan_result gc_plan_typed_impl(const storage_gc_plan_request &request) {
-  if (!request.dry_run) {
-    throw std::invalid_argument("storage_gc_requires_dry_run");
-  }
-  const auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-  const auto referenced = referenced_payload_hashes(request.runtime_dir, request.source_id);
-  const auto payloads = provider->all_payloads();
-  storage_gc_plan_result result{};
-  result.scope = request.source_id.empty() ? "all" : "source";
-  if (!request.source_id.empty()) {
-    result.source_id = request.source_id;
-  }
-  result.payloads_scanned = payloads.size();
-  result.referenced_payloads = referenced.size();
-  for (const auto &payload : payloads) {
-    const auto digest = payload.digest;
-    if (std::find(referenced.begin(), referenced.end(), digest) != referenced.end()) {
-      continue;
-    }
-    result.candidate_bytes += payload.bytes;
-    result.candidates.push_back({digest, payload.uri, payload.bytes, request.source_id.empty()});
-  }
-  result.notes = {"No payloads were deleted.",
-                  request.source_id.empty()
-                      ? "All-scope candidates are unreferenced by retained storage manifests."
-                      : "Source scope candidates are not globally safe to delete because the interim payload store "
-                        "is shared."};
-  return result;
-}
-
-nlohmann::json gc_plan_impl(const storage_service_options &options) {
-  return render_storage_gc_plan_result(default_storage_service().gc_plan(parse_storage_gc_plan_request(options)));
-}
-
-storage_compact_plan_result compact_plan_typed_impl(const storage_compact_plan_request &request) {
-  if (!request.dry_run) {
-    throw std::invalid_argument("storage_compact_requires_dry_run");
-  }
-  storage_compact_plan_result result{};
-  result.scope = request.source_id.empty() ? "all" : "source";
-  if (!request.source_id.empty()) {
-    result.source_id = request.source_id;
-  }
-  result.rebuild_index = rebuild_index_typed_impl({request.runtime_dir, request.source_id, true});
-  result.gc = gc_plan_typed_impl({request.runtime_dir, request.provider, request.source_id, true});
-
-  const auto catalog = catalog_store(request.runtime_dir).read_typed_records();
-  std::map<uint64_t, std::vector<const yijinjing::types::ImportManifestAccepted *>> manifests_by_source;
-  for (const auto &manifest : catalog.manifests) {
-    manifests_by_source[manifest.source_uid].push_back(&manifest);
-  }
-  for (const auto &[source_uid, manifests] : manifests_by_source) {
-    (void)source_uid;
-    const auto current_source_id = fixed_string(manifests.back()->source_id);
-    if (!request.source_id.empty() && current_source_id != request.source_id) {
-      continue;
-    }
-    for (const auto *manifest : manifests) {
-      result.retained_manifests.push_back(
-          {current_source_id,
-           fixed_string(manifest->manifest_id),
-           manifest->entry_count,
-           {fixed_string(manifest->sync_root_algo), fixed_string(manifest->sync_root_value)}});
-    }
-  }
-  result.projection_compact = {PROJECTION_MANIFEST_CATALOG,
-                               manifest_catalog_projection(request.runtime_dir).sqlite_path(), "rebuild-and-vacuum",
-                               true, true};
-  const auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-  result.unsupported = {{"history-archive", "archive bundles are not implemented in this slice"},
-                        {"backend-compact", provider->name() == PROVIDER_ROCKSDB
-                                                ? "RocksDB compaction is not destructive-history compact"
-                                                : "the file backend has no backend compaction"}};
-  result.notes = {"No manifests, payloads, journal frames, or projections were rewritten.",
-                  "This is a reviewable compaction plan, not destructive compaction."};
-  result.ok = result.rebuild_index.ok && result.gc.ok;
-  return result;
-}
-
-nlohmann::json compact_plan_impl(const storage_service_options &options) {
-  return render_storage_compact_plan_result(
-      default_storage_service().compact_plan(parse_storage_compact_plan_request(options)));
-}
-
-// Accept one import manifest into the kernel journals: the manifest-catalog
-// records (header + per-entry deltas + channel cursor, entries document
-// committed into the content store) plus the source-registry alignment
-// (register-once, head update, accepted range). Returns the accepted
-// manifest JSON edge.
-nlohmann::json accept_storage_manifest_impl(const std::string &runtime_dir, const nlohmann::json &input) {
-  auto parsed = parse_storage_export_bundle({{"manifest", input}});
-  (void)default_storage_service().import_bundle({runtime_dir, {}, std::move(parsed), false});
-  const auto provider = provider_cache::instance().acquire(runtime_dir, {});
-  const auto source_id = text_or(input, "source_id", text_or(input, "storage_source_id", "local"));
-  const auto accepted = catalog_store(runtime_dir).latest_manifest_typed(source_id, provider->content_store());
-  if (!accepted.has_value()) {
-    throw std::runtime_error("accepted manifest not found");
-  }
-  return render_manifest_document(*accepted);
-}
-
-nlohmann::json render_manifest_entry_view(const yy_storage::manifest_entry_view &entry) {
-  nlohmann::json rendered = {{"kind", entry.kind},
-                             {"source_id", entry.source_id},
-                             {"source_path", entry.source_path},
-                             {"source_time", entry.source_time},
-                             {"schema_version", entry.schema_version},
-                             {"content_type", entry.content_type},
-                             {"payload_hash", entry.payload_hash},
-                             {"byte_len", entry.byte_len},
-                             {"payload_state", payload_state_text(entry.payload_state)}};
-  if (entry.action_json.has_value()) {
-    rendered["action"] = nlohmann::json::parse(*entry.action_json);
-  }
-  return rendered;
-}
-
-nlohmann::json render_sync_root(const yy_storage::manifest_sync_root_view &root) {
-  return {{"schema", yy_storage::SYNC_ROOT_SCHEMA_V1},
-          {"scope", yy_storage::SYNC_ROOT_SCOPE_ATLAS_IMPORT_MANIFEST},
-          {"proof", yy_storage::SYNC_ROOT_PROOF_LINEAR_CHAIN_V1},
-          {"algorithm", root.algorithm},
-          {"value", root.value},
-          {"entry_count", root.entry_count},
-          {"initial", yy_storage::SYNC_ROOT_INITIAL_SHA256},
-          {"ordering",
-           {{"policy", yy_storage::SYNC_ROOT_ORDERING_POLICY_MANIFEST_ENTRY_SORT_V1},
-            {"fields", {"kind", "source_id", "source_path"}}}}};
-}
-
-nlohmann::json render_manifest_document(const yy_storage::manifest_document_view &manifest) {
-  nlohmann::json entries = nlohmann::json::array();
-  for (const auto &entry : manifest.entries) {
-    entries.push_back(render_manifest_entry_view(entry));
-  }
-  auto range = nlohmann::json::object();
-  if (!manifest.range_since.empty()) {
-    range["since"] = manifest.range_since;
-  }
-  if (!manifest.range_until.empty()) {
-    range["until"] = manifest.range_until;
-  }
-  const auto sync_root = render_sync_root(manifest.sync_root);
-  auto source =
-      nlohmann::json{{"schema", yy_storage::STORAGE_SOURCE_RECORD_SCHEMA_V1},
-                     {"source_id", manifest.source_id},
-                     {"type", manifest.source_type},
-                     {"kind", manifest.source_type == "atlas" ? "adapter" : "local"},
-                     {"coordinate", manifest.source_coordinate},
-                     {"current_head",
-                      {{"head", manifest.source_head}, {"range", range}, {"inventory_hash", manifest.sync_root.value}}},
-                     {"last_manifest_id", manifest.manifest_id},
-                     {"updated_at", ""}};
-  auto accepted = nlohmann::json{{"schema", yy_storage::STORAGE_ACCEPTED_RANGE_SCHEMA_V1},
-                                 {"source_id", manifest.source_id},
-                                 {"manifest_id", manifest.manifest_id},
-                                 {"range", range},
-                                 {"source_head", manifest.source_head},
-                                 {"sync_root", sync_root},
-                                 {"entry_count", manifest.entries.size()},
-                                 {"status", "ok"}};
-  source["accepted_ranges"] = nlohmann::json::array({accepted});
-  return {{"schema", yy_storage::STORAGE_IMPORT_MANIFEST_SCHEMA_V1},
-          {"authority", "yijinjing-journal"},
-          {"manifest_id", manifest.manifest_id},
-          {"scope", manifest.scope},
-          {"source", source},
-          {"source_id", manifest.source_id},
-          {"source_type", manifest.source_type},
-          {"source_head", manifest.source_head},
-          {"range", range},
-          {"counts", {{"records", manifest.entries.size()}}},
-          {"entries", entries},
-          {"payload_inventory", yy_storage::build_storage_payload_inventory(entries)},
-          {"schema_inventory", yy_storage::build_storage_schema_inventory(entries)},
-          {"accepted_ranges", nlohmann::json::array({accepted})},
-          {"sync_root", sync_root},
-          {"idempotency_key", manifest.source_id + ":" + manifest.manifest_id}};
-}
-
-nlohmann::json render_storage_export_bundle_result(const storage_export_bundle_result &result) {
-  const auto manifest = render_manifest_document(result.manifest);
-  nlohmann::json records = nlohmann::json::array();
-  for (const auto &record : result.records) {
-    auto row = render_manifest_entry_view(record.entry);
-    row["scope"] = result.manifest.scope;
-    row["manifest_id"] = result.manifest.manifest_id;
-    row["storage_source_id"] = result.manifest.source_id;
-    row["source_type"] = result.manifest.source_type;
-    row["source_head"] = result.manifest.source_head;
-    row["payload"] =
-        record.payload_json.has_value() ? nlohmann::json::parse(*record.payload_json) : nlohmann::json(nullptr);
-    records.push_back(std::move(row));
-  }
-  auto payload_inventory = yy_storage::build_storage_payload_inventory(manifest.at("entries"));
-  payload_inventory["exported_records"] = records.size();
-  return {{"schema", yy_storage::STORAGE_EXPORT_BUNDLE_SCHEMA_V1},
-          {"bundle_id", result.bundle_id},
-          {"source_id", result.source_id},
-          {"manifest", manifest},
-          {"records", records},
-          {"payload_inventory", payload_inventory},
-          {"schema_inventory", manifest.at("schema_inventory")},
-          {"accepted_ranges", manifest.at("accepted_ranges")},
-          {"sync_root", manifest.at("sync_root")}};
-}
-
-storage_export_bundle_result export_bundle_typed_impl(const storage_export_bundle_request &request) {
-  const auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-  auto manifest =
-      catalog_store(request.runtime_dir).latest_manifest_typed(request.source_id, provider->content_store());
-  if (!manifest.has_value()) {
-    throw std::runtime_error("manifest not found: " + request.source_id);
-  }
-  if (!request.range.since.empty() || !request.range.until.empty()) {
-    std::vector<yy_storage::manifest_entry_view> filtered;
-    for (const auto &entry : manifest->entries) {
-      if (entry.source_time.empty()) {
-        continue;
-      }
-      if (!request.range.since.empty() && entry.source_time < request.range.since) {
-        continue;
-      }
-      if (!request.range.until.empty() && entry.source_time > request.range.until) {
-        continue;
-      }
-      filtered.push_back(entry);
-    }
-    manifest->entries = std::move(filtered);
-    manifest->range_since = request.range.since;
-    manifest->range_until = request.range.until;
-    manifest->sync_root = yy_storage::compute_manifest_sync_root(manifest->entries);
-  }
-
-  storage_export_bundle_result result{};
-  result.bundle_id = manifest->source_id + ":" + manifest->manifest_id;
-  result.source_id = manifest->source_id;
-  result.manifest = *manifest;
-  for (const auto &entry : manifest->entries) {
-    storage_export_record_view record{};
-    record.entry = entry;
-    const auto withheld = entry.payload_state == yy_enums::PayloadState::Redacted ||
-                          entry.payload_state == yy_enums::PayloadState::Absent;
-    if (!withheld && !entry.payload_hash.empty() && provider->payload_exists(entry.payload_hash)) {
-      const auto raw = provider->read_payload(entry.payload_hash);
-      const auto error = yy_storage::verify_payload_ref(raw, entry.payload_hash, entry.byte_len);
-      if (error.empty()) {
-        record.payload_json = raw;
-      } else if (entry.payload_state == yy_enums::PayloadState::Present) {
-        throw std::runtime_error(error + ": " + entry.kind + ":" + entry.source_id);
-      }
-    } else if (!withheld && entry.payload_state == yy_enums::PayloadState::Present) {
-      throw std::runtime_error("payload_missing: " + entry.kind + ":" + entry.source_id);
-    }
-    result.records.push_back(std::move(record));
-  }
-  std::sort(result.records.begin(), result.records.end(), [](const auto &lhs, const auto &rhs) {
-    return std::tie(lhs.entry.kind, lhs.entry.source_id, lhs.entry.source_path) <
-           std::tie(rhs.entry.kind, rhs.entry.source_id, rhs.entry.source_path);
-  });
-  if (request.record_receipt) {
-    catalog_store(request.runtime_dir)
-        .record_export_typed(result.manifest, result.records.size(), request.range.since, request.range.until);
-  }
-  return result;
-}
-
-nlohmann::json export_bundle_generic_impl(const storage_service_options &options, bool record_receipt) {
-  return render_storage_export_bundle_result(
-      default_storage_service().export_bundle({options.runtime_dir,
-                                               options.provider,
-                                               options.source_id,
-                                               {text_or(options.range, "since"), text_or(options.range, "until")},
-                                               record_receipt}));
-}
-
-yy_enums::PayloadState parse_payload_state_text(const std::string &state) {
-  if (state == PAYLOAD_STATE_PRESENT) {
-    return yy_enums::PayloadState::Present;
-  }
-  if (state == PAYLOAD_STATE_REDACTED) {
-    return yy_enums::PayloadState::Redacted;
-  }
-  if (state == PAYLOAD_STATE_ABSENT) {
-    return yy_enums::PayloadState::Absent;
-  }
-  return yy_enums::PayloadState::Missing;
-}
-
-yy_storage::manifest_entry_view parse_manifest_entry_view(const nlohmann::json &entry) {
-  yy_storage::manifest_entry_view parsed{};
-  parsed.kind = text_or(entry, "kind");
-  parsed.source_id = text_or(entry, "source_id");
-  parsed.source_path = text_or(entry, "source_path");
-  parsed.source_time = text_or(entry, "source_time");
-  parsed.schema_version = entry.value("schema_version", uint32_t{0});
-  parsed.content_type = text_or(entry, "content_type");
-  parsed.payload_hash = text_or(entry, "payload_hash");
-  parsed.byte_len = entry.value("byte_len", uint64_t{0});
-  parsed.payload_state = parse_payload_state_text(text_or(entry, "payload_state", "missing"));
-  if (entry.contains("action") && entry.at("action").is_object()) {
-    parsed.action_json = canonical_json(entry.at("action"));
-  }
-  return parsed;
-}
-
-storage_export_bundle_result parse_storage_export_bundle(const nlohmann::json &bundle) {
-  if (!bundle.is_object()) {
-    throw std::invalid_argument("bundle_manifest_missing");
-  }
-  const auto manifest_edge = object_or_empty(bundle, "manifest");
-  if (manifest_edge.empty()) {
-    throw std::invalid_argument("bundle_manifest_missing");
-  }
-  storage_export_bundle_result parsed{};
-  parsed.bundle_id = text_or(bundle, "bundle_id");
-  parsed.source_id = text_or(bundle, "source_id");
-  auto &manifest = parsed.manifest;
-  manifest.manifest_id = text_or(manifest_edge, "manifest_id");
-  manifest.scope = text_or(manifest_edge, "scope");
-  manifest.source_id = text_or(manifest_edge, "source_id", text_or(manifest_edge, "storage_source_id", "local"));
-  manifest.source_type = text_or(manifest_edge, "source_type");
-  manifest.source_head = text_or(manifest_edge, "source_head");
-  manifest.source_coordinate =
-      text_or(object_or_empty(manifest_edge, "source"), "coordinate", text_or(manifest_edge, "source_coordinate"));
-  const auto range = object_or_empty(manifest_edge, "range");
-  manifest.range_since = text_or(range, "since");
-  manifest.range_until = text_or(range, "until");
-  for (const auto &entry : array_or_empty(manifest_edge, "entries")) {
-    if (!entry.is_object()) {
-      throw std::invalid_argument("bundle_manifest_invalid: manifest_entries_invalid");
-    }
-    manifest.entries.push_back(parse_manifest_entry_view(entry));
-  }
-  const auto root = object_or_empty(manifest_edge, "sync_root");
-  manifest.sync_root = {text_or(root, "algorithm"), text_or(root, "value"), root.value("entry_count", uint64_t{0})};
-  for (const auto &record : array_or_empty(bundle, "records")) {
-    if (!record.is_object()) {
-      continue;
-    }
-    storage_export_record_view row{};
-    row.entry = parse_manifest_entry_view(record);
-    if (record.contains("payload") && !record.at("payload").is_null()) {
-      row.payload_json = canonical_json(record.at("payload"));
-    }
-    parsed.records.push_back(std::move(row));
-  }
-  return parsed;
-}
-
-storage_import_bundle_result import_bundle_typed_impl(const storage_import_bundle_request &request) {
-  const auto &manifest = request.bundle.manifest;
-  if (manifest.manifest_id.empty())
-    throw std::invalid_argument("bundle_manifest_invalid: missing_field: manifest_id");
-  if (manifest.source_id.empty())
-    throw std::invalid_argument("bundle_manifest_invalid: missing_field: source_id");
-  if (request.verify) {
-    const auto expected = yy_storage::compute_manifest_sync_root(manifest.entries);
-    if (expected.algorithm != manifest.sync_root.algorithm || expected.value != manifest.sync_root.value ||
-        expected.entry_count != manifest.sync_root.entry_count) {
-      throw std::invalid_argument("bundle_manifest_invalid: sync_root_mismatch");
-    }
-  }
-  const auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-  for (const auto &record : request.bundle.records) {
-    if (!record.payload_json.has_value()) {
-      continue;
-    }
-    if (record.entry.payload_state == yy_enums::PayloadState::Redacted ||
-        record.entry.payload_state == yy_enums::PayloadState::Absent) {
-      continue;
-    }
-    auto digest = record.entry.payload_hash;
-    if (digest.empty()) {
-      digest = yy_storage::compute_content_hash_value(*record.payload_json, yy_storage::CONTENT_HASH_ALGORITHM_SHA256);
-    }
-    const auto error = yy_storage::verify_payload_ref(*record.payload_json, digest, record.payload_json->size());
-    if (!error.empty()) {
-      throw std::invalid_argument("storage_payload_invalid: " + error);
-    }
-    provider->write_payload(digest, *record.payload_json);
-  }
-
-  const auto accepted = catalog_store(request.runtime_dir).accept_manifest_typed(manifest, provider->content_store());
-  const auto registry = registry_store(request.runtime_dir);
-  bool registered = false;
-  for (const auto &[source_uid, source] : registry.fold_typed_records().sources) {
-    (void)source_uid;
-    if (source.registered && fixed_string(source.registration.source_id) == accepted.source_id) {
-      registered = true;
-      break;
-    }
-  }
-  if (!registered) {
-    yy_storage::source_register_options reg{};
-    reg.source_id = accepted.source_id;
-    reg.kind = accepted.source_type == "atlas" ? yy_enums::SourceKind::Adapter : yy_enums::SourceKind::Local;
-    reg.coordinate = accepted.source_coordinate;
-    reg.head = accepted.source_head;
-    (void)registry.register_source(reg);
-  }
-  yy_storage::source_head_update_options head{};
-  head.source_id = accepted.source_id;
-  head.head = accepted.source_head;
-  head.inventory_hash_algo = accepted.sync_root.algorithm;
-  head.inventory_hash = accepted.sync_root.value;
-  (void)registry.update_head(head);
-  yy_storage::accepted_range_options range{};
-  range.source_id = accepted.source_id;
-  range.manifest_id = accepted.manifest_id;
-  (void)registry.record_accepted_range(range);
-  return {true, "source", accepted.source_id, accepted.manifest_id, request.bundle.records.size()};
-}
-
-nlohmann::json render_storage_import_bundle_result(const storage_import_bundle_result &result) {
-  return {{"ok", result.ok},
-          {"scope", result.scope},
-          {"source_id", result.source_id},
-          {"manifest_id", result.manifest_id},
-          {"records", result.records}};
-}
-
-storage_verify_sync_result verify_sync_typed_impl(const storage_verify_sync_request &request) {
-  storage_verify_sync_result result{};
-  result.source_id = request.source_id;
-  result.source_fsck =
-      default_storage_service().fsck({request.runtime_dir, request.provider, request.provider_config_source,
-                                      storage_fsck_scope::Source, request.source_id, 0, false});
-  if (!result.source_fsck.ok) {
-    result.ok = false;
-    return result;
-  }
-
-  const auto bundle =
-      default_storage_service().export_bundle({request.runtime_dir, request.provider, request.source_id, {}, false});
-  result.exported_records = bundle.records.size();
-  result.local_sync_root = {bundle.manifest.sync_root.algorithm, bundle.manifest.sync_root.value};
-  const auto temp_root = fs::temp_directory_path() / ("kungfu-storage-sync-" + std::to_string(std::random_device{}()) +
-                                                      "-" + std::to_string(std::random_device{}()));
-  result.imported_runtime_dir = temp_root.string();
-  try {
-    result.import = default_storage_service().import_bundle({temp_root.string(), request.provider, bundle, true});
-    result.imported_fsck =
-        default_storage_service().fsck({temp_root.string(), request.provider, request.provider_config_source,
-                                        storage_fsck_scope::Source, request.source_id, 0, false});
-    const auto imported_provider = provider_cache::instance().acquire(temp_root.string(), request.provider);
-    const auto imported =
-        catalog_store(temp_root.string()).latest_manifest_typed(request.source_id, imported_provider->content_store());
-    if (imported.has_value()) {
-      result.imported_sync_root = {imported->sync_root.algorithm, imported->sync_root.value};
-    }
-    fs::remove_all(temp_root);
-  } catch (...) {
-    fs::remove_all(temp_root);
-    throw;
-  }
-  result.sync_roots_match = result.local_sync_root.algorithm == result.imported_sync_root.algorithm &&
-                            result.local_sync_root.value == result.imported_sync_root.value;
-  result.ok = result.imported_fsck.ok && result.sync_roots_match;
-  return result;
-}
-
-nlohmann::json render_storage_verify_sync_result(const storage_verify_sync_result &result) {
-  if (!result.source_fsck.ok) {
-    return {{"ok", false},
-            {"scope", "source"},
-            {"source_id", result.source_id},
-            {"errors", nlohmann::json::array({{{"code", "source_fsck_failed"},
-                                               {"fsck", render_storage_fsck_result(result.source_fsck)}}})}};
-  }
-  auto imported_fsck = render_storage_fsck_result(result.imported_fsck);
-  imported_fsck = replace_string_subtree(imported_fsck, result.imported_runtime_dir, "<sync-runtime>");
-  return {{"ok", result.ok},
-          {"scope", result.scope},
-          {"source_id", result.source_id},
-          {"exported_records", result.exported_records},
-          {"import", render_storage_import_bundle_result(result.import)},
-          {"local_sync_root",
-           render_sync_root({result.local_sync_root.algorithm, result.local_sync_root.value, result.exported_records})},
-          {"imported_sync_root", render_sync_root({result.imported_sync_root.algorithm, result.imported_sync_root.value,
-                                                   result.exported_records})},
-          {"sync_roots_match", result.sync_roots_match},
-          {"source_fsck", render_storage_fsck_result(result.source_fsck)},
-          {"imported_fsck", imported_fsck}};
-}
-
-class file_storage_json_edge_service {
-public:
-  [[nodiscard]] nlohmann::json status(const storage_service_options &options) const {
-    return render_storage_status_result(default_storage_service().status(parse_storage_status_request(options)));
-  }
-
-  [[nodiscard]] nlohmann::json fsck(const storage_service_options &options) const { return fsck_impl(options); }
-
-  [[nodiscard]] nlohmann::json repair_plan(const storage_service_options &options) const {
-    return repair_plan_impl(options);
-  }
-
-  [[nodiscard]] nlohmann::json repair_fetch(const storage_service_options &options) const {
-    return repair_fetch_impl(options);
-  }
-
-  [[nodiscard]] nlohmann::json repair_apply(const storage_service_options &options) const {
-    return repair_apply_impl(options);
-  }
-
-  [[nodiscard]] nlohmann::json export_bundle(const storage_service_options &options) const {
-    if (options.scope == "episode") {
-      return episode_export_bundle_impl(options);
-    }
-    // The public export operation records the export-bundle receipt; internal
-    // read-only exports (verify_sync round trips, repair-fetch evidence scans
-    // over mirror runtimes) call the impl without the receipt.
-    return export_bundle_generic_impl(options, /*record_receipt=*/true);
-  }
-
-  [[nodiscard]] nlohmann::json import_bundle(const storage_service_options &options) const {
-    if (!options.bundle.is_object()) {
-      throw std::invalid_argument("bundle_manifest_missing");
-    }
-    if (options.scope == "episode" || text_or(options.bundle, "schema") == "kungfu.storage.episode-bundle/v1") {
-      return episode_import_bundle_impl(options);
-    }
-    return render_storage_import_bundle_result(default_storage_service().import_bundle(
-        {options.runtime_dir, options.provider, parse_storage_export_bundle(options.bundle), options.verify}));
-  }
-
-  [[nodiscard]] nlohmann::json rebuild_index(const storage_service_options &options) const {
-    return rebuild_index_impl(options);
-  }
-
-  [[nodiscard]] nlohmann::json gc_plan(const storage_service_options &options) const { return gc_plan_impl(options); }
-
-  [[nodiscard]] nlohmann::json compact_plan(const storage_service_options &options) const {
-    return compact_plan_impl(options);
-  }
-
-  [[nodiscard]] nlohmann::json verify_sync(const storage_service_options &options) const {
-    return render_storage_verify_sync_result(default_storage_service().verify_sync(
-        {options.runtime_dir, options.provider, options.provider_config_source, options.source_id}));
-  }
-
-  [[nodiscard]] nlohmann::json layout(const storage_service_options &options) const {
-    storage_layout_request request{};
-    request.runtime_dir = options.runtime_dir;
-    request.runtime_home = text_or(options.operation_options, "runtime_home");
-    request.config_home = text_or(options.operation_options, "config_home");
-    request.provider = options.provider;
-    return workspace_episode_layout_json(default_storage_service().layout(request));
-  }
-
-  [[nodiscard]] nlohmann::json query_plan(const storage_service_options &options) const {
-    const auto action = text_or(options.operation_options, "action", "validate");
-    if (action == "capabilities") {
-      return query::query_capabilities_json();
-    }
-    if (action == "schema") {
-      return query::query_definition_schema_json();
-    }
-    if (action == "describe") {
-      return query::query_object_description_json(text_or(options.operation_options, "object", "episodes"));
-    }
-    if (action == "examples") {
-      return query::query_examples_json();
-    }
-
-    auto definition = query::parse_query_definition(options.query_definition);
-    if (action == "compile-sql") {
-      if (options.query.empty()) {
-        throw std::invalid_argument("compile-sql requires a non-empty SQL query");
-      }
-      definition = query::compile_episode_sql(options.query, definition);
-    }
-    const auto plan = query::plan_query(definition);
-    if (action == "compile-sql") {
-      return {{"schema", "kungfu.query.sql-compilation/v1"},
-              {"sql", options.query},
-              {"definition", query::query_definition_json(definition)},
-              {"logical_plan", query::logical_plan_json(plan)},
-              {"query_definition_hash", plan.query_definition_hash},
-              {"logical_plan_hash", plan.logical_plan_hash}};
-    }
-    if (action == "validate") {
-      return {{"schema", query::QUERY_VALIDATION_SCHEMA_V1},
-              {"ok", true},
-              {"definition", query::query_definition_json(definition)},
-              {"query_definition_hash", plan.query_definition_hash},
-              {"logical_plan_hash", plan.logical_plan_hash}};
-    }
-    if (action == "explain") {
-      const auto engine = text_or(options.operation_options, "engine", "authority");
-      if (engine != "authority" && engine != "sqlite") {
-        throw std::invalid_argument("query engine must be authority or sqlite");
-      }
-      const auto authority_engine =
-          definition.object == "fact-state" ? "fact-authority-scan/v1" : "episode-authority-scan/v1";
-      if (definition.object == "fact-state" && engine == "sqlite") {
-        throw std::invalid_argument("object=fact-state currently supports only the authority engine");
-      }
-      return {{"schema", query::QUERY_EXPLAIN_SCHEMA_V1},
-              {"definition", query::query_definition_json(definition)},
-              {"logical_plan", query::logical_plan_json(plan)},
-              {"physical",
-               {{"engine", engine == "authority" ? authority_engine : "episode-sqlite-projection/v1"},
-                {"bounded", true},
-                {"limit", definition.limit},
-                {"cost",
-                 {{"class", engine == "authority" ? "bounded-authority-scan" : "bounded-sqlite-projection"},
-                  {"upper_bound_rows", definition.limit},
-                  {"authority_records", "runtime-dependent"}}}}}};
-    }
-    throw std::invalid_argument("unsupported query planner action: " + action);
-  }
-
-  [[nodiscard]] nlohmann::json fact_query(const storage_service_options &options) const {
-    const auto definition = query::parse_query_definition(options.query_definition);
-    const auto plan = query::plan_query(definition);
-    const auto engine = text_or(options.operation_options, "engine", "authority");
-    if (engine == "authority") {
-      if (definition.object == "fact-state") {
-        return query::query_result_json(query::run_fact_state_authority_scan(options.runtime_dir, plan));
-      }
-      return query::query_result_json(query::run_episode_authority_scan(options.runtime_dir, plan));
-    }
-    if (engine == "sqlite") {
-      if (definition.object != "episodes") {
-        throw std::invalid_argument("object=fact-state currently supports only the authority engine");
-      }
-      return query::query_result_json(query::run_episode_sqlite_projection(options.runtime_dir, plan));
-    }
-    throw std::invalid_argument("query engine must be authority or sqlite");
-  }
-
-  [[nodiscard]] nlohmann::json fact_changelog(const storage_service_options &options) const {
-    const auto definition = query::parse_query_definition(options.query_definition);
-    const auto plan = query::plan_query(definition);
-    const auto resume_token = object_or_empty(options.operation_options, "resume_token");
-    const auto max_messages = uint64_or(options.operation_options, "max_messages", 100);
-    return query::changelog_page_json(
-        query::run_query_changelog(options.runtime_dir, plan, resume_token, max_messages));
-  }
-
-  [[nodiscard]] nlohmann::json saved_query_catalog(const storage_service_options &options) const {
-    const auto action = text_or(options.operation_options, "action", "list");
-    if (action == "contract") {
-      return query::saved_query_catalog_contract();
-    }
-    if (action == "put") {
-      return query::saved_query_put(options.runtime_dir, object_or_empty(options.operation_options, "saved_view"),
-                                    text_or(options.operation_options, "query_id"),
-                                    uint64_or(options.operation_options, "expected_revision"),
-                                    int64_or(options.operation_options, "system_time"));
-    }
-    if (action == "get") {
-      return query::saved_query_get(options.runtime_dir, text_or(options.operation_options, "query_id"),
-                                    bool_or(options.operation_options, "include_deleted", false));
-    }
-    if (action == "list") {
-      return query::saved_query_list(options.runtime_dir, bool_or(options.operation_options, "include_deleted", false));
-    }
-    if (action == "history") {
-      return query::saved_query_history(options.runtime_dir, text_or(options.operation_options, "query_id"));
-    }
-    if (action == "delete") {
-      return query::saved_query_delete(options.runtime_dir, text_or(options.operation_options, "query_id"),
-                                       uint64_or(options.operation_options, "expected_revision"),
-                                       int64_or(options.operation_options, "system_time"));
-    }
-    if (action == "rebuild") {
-      return query::saved_query_rebuild(options.runtime_dir);
-    }
-    throw std::invalid_argument("unsupported saved-query catalog action: " + action);
-  }
-
-  [[nodiscard]] nlohmann::json profile_lifecycle(const storage_service_options &options) const {
-    const auto action = text_or(options.operation_options, "action", "list");
-    if (action == "contract") {
-      return profile::profile_lifecycle_contract();
-    }
-    if (action == "inspect") {
-      return profile::inspect_profile(text_or(options.operation_options, "profile_path"),
-                                      object_or_empty(options.operation_options, "member_roots"));
-    }
-    if (action == "plan") {
-      return profile::plan_profile_lifecycle(options.runtime_dir,
-                                             object_or_empty(options.operation_options, "request"));
-    }
-    if (action == "apply") {
-      return profile::apply_profile_lifecycle(options.runtime_dir, object_or_empty(options.operation_options, "plan"),
-                                              text_or(options.operation_options, "authorization_id"),
-                                              int64_or(options.operation_options, "system_time"));
-    }
-    if (action == "get") {
-      return profile::get_profile(options.runtime_dir, text_or(options.operation_options, "profile_id"),
-                                  bool_or(options.operation_options, "include_removed", false),
-                                  int64_or(options.operation_options, "cut_system_time"));
-    }
-    if (action == "list") {
-      return profile::list_profiles(options.runtime_dir, bool_or(options.operation_options, "include_removed", false),
-                                    int64_or(options.operation_options, "cut_system_time"));
-    }
-    if (action == "history") {
-      return profile::profile_history(options.runtime_dir, text_or(options.operation_options, "profile_id"));
-    }
-    throw std::invalid_argument("unsupported Profile lifecycle action: " + action);
-  }
-
-  [[nodiscard]] nlohmann::json kfx_runtime(const storage_service_options &options) const {
-    const auto action = text_or(options.operation_options, "action", "contract");
-    if (action == "contract") {
-      return kfx::native_kfx_contract();
-    }
-    if (action == "validate") {
-      return kfx::validate_native_kfx_document(text_or(options.operation_options, "kind"),
-                                               object_or_empty(options.operation_options, "document"));
-    }
-    if (action == "list" || action == "inspect" || action == "resolve" || action == "plan" || action == "status") {
-      return kfx::query_native_kfx_registry(action, object_or_empty(options.operation_options, "request"));
-    }
-    throw std::invalid_argument("unsupported native KFX runtime action: " + action);
-  }
-
-  [[nodiscard]] nlohmann::json fact_contract(const storage_service_options &options) const {
-    (void)options;
-    return facts::fact_contract_json();
-  }
-
-  [[nodiscard]] nlohmann::json fact_declare_world(const storage_service_options &options) const {
-    return facts::declare_contract_world(options.runtime_dir, object_or_empty(options.operation_options, "declaration"),
-                                         int64_or(options.operation_options, "system_time"));
-  }
-
-  [[nodiscard]] nlohmann::json fact_declare_surface(const storage_service_options &options) const {
-    return facts::declare_fact_surface(options.runtime_dir, object_or_empty(options.operation_options, "declaration"),
-                                       int64_or(options.operation_options, "system_time"));
-  }
-
-  [[nodiscard]] nlohmann::json fact_observe(const storage_service_options &options) const {
-    return facts::record_observation(options.runtime_dir, object_or_empty(options.operation_options, "observation"),
-                                     int64_or(options.operation_options, "system_time"));
-  }
-
-  [[nodiscard]] nlohmann::json fact_state(const storage_service_options &options) const {
-    return facts::query_fact_state(options.runtime_dir, int64_or(options.operation_options, "cut_system_time"),
-                                   text_or(options.operation_options, "subject_key"));
-  }
-
-  [[nodiscard]] nlohmann::json fact_library_contract(const storage_service_options &options) const {
-    (void)options;
-    return {
-        {"schema", "kungfu.facts.library-contract/v1"},
-        {"owner", "libkungfu"},
-        {"authority", "yijinjing-journal"},
-        {"content_store", {{"schemas", "schemas"}, {"payloads", "payloads"}}},
-        {"schema_validation",
-         {{"profile", "json-schema-object-subset/v1"},
-          {"keywords", nlohmann::json::array({"type", "properties", "required", "items", "additionalProperties"})}}},
-        {"semantic_profile",
-         {{"id", "declared-facts-v1"},
-          {"identity_policy", "subject-key/v1"},
-          {"valid_time_policy", "explicit-range/v1"},
-          {"system_time_policy", "journal-event-time/v1"},
-          {"causal_time_policy", "event-parent/v1"},
-          {"reducer_policy", "latest-admitted-per-source/v1"},
-          {"correction_policy", "explicit-target/v1"},
-          {"retraction_policy", "explicit-target/v1"},
-          {"conflict_policy", "preserve-source-claims/v1"},
-          {"redaction_policy", "hash-and-ref/v1"},
-          {"compatibility_policy", "exact-schema-root/v1"}}},
-        {"operations", nlohmann::json::array({"type-create", "type-list", "material-put", "material-list"})},
-        {"portability", {{"full", "owned schema and payload bytes"}, {"thin", "declared refs only"}}},
-        {"known_limits", nlohmann::json::array({"one fixed semantic profile", "bounded JSON Schema object subset",
-                                                "no mutable global-latest workspace binding"})}};
-  }
-
-  [[nodiscard]] nlohmann::json fact_type_list(const storage_service_options &options) const {
-    const auto state =
-        facts::query_fact_state(options.runtime_dir, int64_or(options.operation_options, "cut_system_time"));
-    const auto catalog = object_or_empty(state, "catalog");
-    return {{"schema", "kungfu.facts.type-catalog/v1"},
-            {"scope", text_or(options.operation_options, "scope", "selected-data-root")},
-            {"semantic_profile", "declared-facts-v1"},
-            {"contract_worlds", array_or_empty(catalog, "contract_worlds")},
-            {"fact_types", array_or_empty(catalog, "fact_surfaces")},
-            {"proof", state.at("proof")}};
-  }
-
-  [[nodiscard]] nlohmann::json fact_type_create(const storage_service_options &options) const {
-    const auto definition = object_or_empty(options.operation_options, "definition");
-    const auto type_id = required_text(definition, "id");
-    const auto version = required_text(definition, "version");
-    const auto sources = array_or_empty(definition, "source_authorities");
-    if (sources.empty()) {
-      throw std::invalid_argument("source_authorities requires at least one source id");
-    }
-    const auto schema = object_or_empty(definition, "schema");
-    if (schema.empty()) {
-      throw std::invalid_argument("schema must be a non-empty JSON object");
-    }
-    if (text_or(schema, "type") != "object") {
-      throw std::invalid_argument("managed fact schema must use the supported top-level object profile");
-    }
-    const auto raw_schema = canonical_json(schema);
-    const auto schema_hash = yy_storage::format_content_hash(yy_storage::compute_content_hash(raw_schema));
-    const auto effective_from =
-        int64_or(definition, "effective_from",
-                 int64_or(options.operation_options, "system_time", kungfu::yijinjing::time::now_in_nano()));
-    const auto effective_until = int64_or(definition, "effective_until");
-    const auto world_id = text_or(definition, "contract_world_id", type_id + ".world");
-
-    const auto existing = fact_type_list(options);
-    const auto provider = shared_provider(options);
-    nlohmann::json world_reference;
-    for (const auto &world : array_or_empty(existing, "contract_worlds")) {
-      if (text_or(world, "id") == world_id && text_or(world, "version") == version) {
-        world_reference = {{"id", world.at("id")}, {"version", world.at("version")}, {"root", world.at("root")}};
-        break;
-      }
-    }
-    for (const auto &type : array_or_empty(existing, "fact_types")) {
-      if (text_or(type, "id") == type_id && text_or(type, "version") == version) {
-        if (text_or(type, "schema_owner_root") != schema_hash) {
-          throw std::invalid_argument("fact type version already exists with a different schema root");
-        }
-        if (canonical_json(array_or_empty(type, "source_authorities")) != canonical_json(sources) ||
-            text_or(object_or_empty(type, "contract_world"), "id") != world_id) {
-          throw std::invalid_argument("fact type version already exists with a different immutable definition");
-        }
-        const auto stored =
-            provider->content_store().put_if_absent("schemas", raw_schema, yy_storage::parse_content_hash(schema_hash));
-        if (!stored.ok()) {
-          throw std::runtime_error("fact schema store failed: " + stored.message);
-        }
-        return {{"schema", "kungfu.facts.type-write/v1"},
-                {"ok", true},
-                {"status", "already_present"},
-                {"definition", type},
-                {"schema_hash", schema_hash},
-                {"schema_store", content_result_json(stored)}};
-      }
-    }
-
-    const auto stored =
-        provider->content_store().put_if_absent("schemas", raw_schema, yy_storage::parse_content_hash(schema_hash));
-    if (!stored.ok()) {
-      throw std::runtime_error("fact schema store failed: " + stored.message);
-    }
-    nlohmann::json world_receipt = nullptr;
-    if (world_reference.is_null() || world_reference.empty()) {
-      world_receipt = facts::declare_contract_world(options.runtime_dir,
-                                                    {{"id", world_id},
-                                                     {"version", version},
-                                                     {"effective_from", effective_from},
-                                                     {"effective_until", effective_until},
-                                                     {"fact_surface_ids", nlohmann::json::array({type_id})}},
-                                                    effective_from);
-      world_reference = world_receipt.at("reference");
-    }
-    nlohmann::json surface = {{"id", type_id},
-                              {"version", version},
-                              {"contract_world", world_reference},
-                              {"effective_from", effective_from},
-                              {"effective_until", effective_until},
-                              {"schema_owner_root", schema_hash},
-                              {"source_authorities", sources},
-                              {"identity_policy", "subject-key/v1"},
-                              {"valid_time_policy", "explicit-range/v1"},
-                              {"system_time_policy", "journal-event-time/v1"},
-                              {"causal_time_policy", "event-parent/v1"},
-                              {"reducer_policy", "latest-admitted-per-source/v1"},
-                              {"correction_policy", "explicit-target/v1"},
-                              {"retraction_policy", "explicit-target/v1"},
-                              {"conflict_policy", "preserve-source-claims/v1"},
-                              {"redaction_policy", "hash-and-ref/v1"},
-                              {"compatibility_policy", "exact-schema-root/v1"},
-                              {"known_limits", nlohmann::json::array({"declared-facts-v1 fixed semantic profile"})}};
-    const auto receipt = facts::declare_fact_surface(options.runtime_dir, surface, effective_from, schema_hash);
-    return {{"schema", "kungfu.facts.type-write/v1"},
-            {"ok", true},
-            {"status", "created"},
-            {"definition", receipt.at("declaration")},
-            {"schema_hash", schema_hash},
-            {"schema_store", content_result_json(stored)},
-            {"world", world_receipt},
-            {"receipt", receipt}};
-  }
-
-  [[nodiscard]] nlohmann::json fact_material_put(const storage_service_options &options) const {
-    const auto material = object_or_empty(options.operation_options, "material");
-    const auto type_id = required_text(material, "type_id");
-    const auto type_version = required_text(material, "type_version");
-    const auto source_id = required_text(material, "source_id");
-    const auto subject_key = required_text(material, "subject_key");
-    if (!material.contains("payload") || !material.at("payload").is_object()) {
-      throw std::invalid_argument("payload must be a JSON object");
-    }
-    const auto types = fact_type_list(options);
-    nlohmann::json selected;
-    for (const auto &type : array_or_empty(types, "fact_types")) {
-      if (text_or(type, "id") == type_id && text_or(type, "version") == type_version) {
-        selected = type;
-        break;
-      }
-    }
-    if (selected.is_null() || selected.empty()) {
-      throw std::invalid_argument("fact type version is not registered in the selected data root");
-    }
-    const auto provider = shared_provider(options);
-    const auto schema_hash = required_text(selected, "schema_owner_root");
-    const auto stored_schema = provider->content_store().get("schemas", yy_storage::parse_content_hash(schema_hash));
-    if (!stored_schema.ok()) {
-      throw std::runtime_error("fact type schema content is unavailable: " + schema_hash);
-    }
-    const auto schema = nlohmann::json::parse(stored_schema.bytes);
-    validate_managed_json_value(schema, material.at("payload"), "$");
-    const auto raw_payload = canonical_json(material.at("payload"));
-    const auto payload_hash = yy_storage::format_content_hash(yy_storage::compute_content_hash(raw_payload));
-    const auto stored =
-        provider->content_store().put_if_absent("payloads", raw_payload, yy_storage::parse_content_hash(payload_hash));
-    if (!stored.ok()) {
-      throw std::runtime_error("fact payload store failed: " + stored.message);
-    }
-    const auto system_time = int64_or(options.operation_options, "system_time", kungfu::yijinjing::time::now_in_nano());
-    auto observation_id = text_or(material, "observation_id");
-    if (observation_id.empty()) {
-      const auto identity = canonical_json({{"type_id", type_id},
-                                            {"type_version", type_version},
-                                            {"source_id", source_id},
-                                            {"subject_key", subject_key},
-                                            {"payload_hash", payload_hash},
-                                            {"system_time", system_time}});
-      const auto identity_hash = yy_storage::compute_content_hash(identity).value;
-      observation_id = "obs-" + identity_hash.substr(0, 24);
-    }
-    const auto world = object_or_empty(selected, "contract_world");
-    nlohmann::json observation = {{"observation_id", observation_id},
-                                  {"contract_world_id", required_text(world, "id")},
-                                  {"fact_surface_id", type_id},
-                                  {"schema_owner_root", required_text(selected, "schema_owner_root")},
-                                  {"source_id", source_id},
-                                  {"subject_key", subject_key},
-                                  {"valid_from", int64_or(material, "valid_from", system_time)},
-                                  {"valid_until", int64_or(material, "valid_until")},
-                                  {"payload_hash", payload_hash},
-                                  {"payload_ref", "content:payloads/" + payload_hash},
-                                  {"action", text_or(material, "action", "assert")},
-                                  {"target_observation_id", text_or(material, "target_observation_id")}};
-    const auto receipt = facts::record_observation(options.runtime_dir, observation, system_time, payload_hash);
-    return {{"schema", "kungfu.facts.material-write/v1"},
-            {"ok", receipt.at("admission").at("outcome") == "admitted"},
-            {"payload_hash", payload_hash},
-            {"payload_store", content_result_json(stored)},
-            {"receipt", receipt}};
-  }
-
-  [[nodiscard]] nlohmann::json fact_material_list(const storage_service_options &options) const {
-    auto state = facts::query_fact_state(options.runtime_dir, int64_or(options.operation_options, "cut_system_time"),
-                                         text_or(options.operation_options, "subject_key"));
-    const auto type_id = text_or(options.operation_options, "type_id");
-    auto filter = [&type_id](nlohmann::json rows) {
-      if (type_id.empty())
-        return rows;
-      nlohmann::json selected = nlohmann::json::array();
-      for (const auto &row : rows) {
-        if (text_or(row, "fact_surface_id") == type_id)
-          selected.push_back(row);
-      }
-      return selected;
-    };
-    state["canonical_facts"] = filter(array_or_empty(state, "canonical_facts"));
-    state["observation_history"] = filter(array_or_empty(state, "observation_history"));
-    nlohmann::json payloads = nlohmann::json::object();
-    const auto provider = shared_provider(options);
-    for (const auto &row : state.at("observation_history")) {
-      const auto hash_text = text_or(row, "payload_hash");
-      if (hash_text.empty() || payloads.contains(hash_text))
-        continue;
-      const auto stored = provider->content_store().get("payloads", yy_storage::parse_content_hash(hash_text));
-      if (stored.ok()) {
-        payloads[hash_text] = nlohmann::json::parse(stored.bytes);
-      }
-    }
-    return {{"schema", "kungfu.facts.material-catalog/v1"},
-            {"type_id", type_id.empty() ? nlohmann::json(nullptr) : nlohmann::json(type_id)},
-            {"state", state},
-            {"payloads", payloads}};
-  }
-
-  [[nodiscard]] nlohmann::json fact_library_export(const storage_service_options &options) const {
-    std::vector<uint64_t> episode_ids;
-    std::unordered_set<uint64_t> seen;
-    const auto push_episode = [&episode_ids, &seen](uint64_t episode_id) {
-      if (episode_id != 0 && seen.insert(episode_id).second)
-        episode_ids.push_back(episode_id);
-    };
-    const auto types = fact_type_list(options);
-    for (const auto &world : array_or_empty(types, "contract_worlds"))
-      push_episode(uint64_or(world, "episode_id"));
-    for (const auto &type : array_or_empty(types, "fact_types"))
-      push_episode(uint64_or(type, "episode_id"));
-    const auto materials = fact_material_list(options);
-    const auto state = object_or_empty(materials, "state");
-    for (const auto &row : array_or_empty(state, "observation_history"))
-      push_episode(uint64_or(row, "episode_id"));
-    const auto assessments = trust::list_assessments(options.runtime_dir);
-    for (const auto &row : array_or_empty(assessments, "assessments")) {
-      push_episode(uint64_or(row, "parent_episode_id"));
-      push_episode(uint64_or(row, "assessment_episode_id"));
-    }
-    nlohmann::json bundles = nlohmann::json::array();
-    uint64_t missing_frame_count = 0;
-    uint64_t missing_ref_payload_count = 0;
-    bool episodes_self_contained = true;
-    for (const auto episode_id : episode_ids) {
-      auto export_options = options;
-      export_options.scope = "episode";
-      export_options.episode_id = episode_id;
-      export_options.operation_options["episode_id"] = episode_id;
-      auto rendered = render_storage_episode_bundle_result(episode_export_bundle_typed_impl(export_options));
-      const auto material = object_or_empty(rendered, "material");
-      missing_frame_count += uint64_or(material, "missing_frame_count");
-      missing_ref_payload_count += uint64_or(material, "missing_ref_payload_count");
-      episodes_self_contained = episodes_self_contained && bool_or(rendered, "self_contained", false);
-      bundles.push_back(std::move(rendered));
-    }
-    const auto thin = bool_or(options.operation_options, "thin", false);
-    const auto self_contained = !thin && episodes_self_contained;
-    return {{"schema", "kungfu.facts.library-bundle/v1"},
-            {"mode", thin ? "thin" : "full"},
-            {"self_contained", self_contained},
-            {"semantic_profile", "declared-facts-v1"},
-            {"episode_count", bundles.size()},
-            {"material",
-             {{"missing_frame_count", missing_frame_count}, {"missing_ref_payload_count", missing_ref_payload_count}}},
-            {"episodes", bundles},
-            {"catalog", types},
-            {"assessments", assessments}};
-  }
-
-  [[nodiscard]] nlohmann::json fact_library_import(const storage_service_options &options) const {
-    const auto bundle = object_or_empty(options.operation_options, "library_bundle");
-    if (text_or(bundle, "schema") != "kungfu.facts.library-bundle/v1") {
-      throw std::invalid_argument("fact_library_bundle_invalid");
-    }
-    const auto dry_run = bool_or(options.operation_options, "dry_run", true);
-    nlohmann::json preflight_receipts = nlohmann::json::array();
-    bool ok = true;
-    for (const auto &episode : array_or_empty(bundle, "episodes")) {
-      auto import_options = options;
-      import_options.scope = "episode";
-      import_options.bundle = episode;
-      import_options.dry_run = true;
-      const auto receipt = episode_import_bundle_impl(import_options);
-      ok = ok && receipt.value("ok", false);
-      preflight_receipts.push_back(receipt);
-    }
-    if (dry_run || !ok) {
-      return {{"schema", "kungfu.facts.library-import/v1"},
-              {"ok", ok},
-              {"dry_run", true},
-              {"source_mode", text_or(bundle, "mode")},
-              {"receipt_count", preflight_receipts.size()},
-              {"receipts", preflight_receipts}};
-    }
-    nlohmann::json receipts = nlohmann::json::array();
-    for (const auto &episode : array_or_empty(bundle, "episodes")) {
-      auto import_options = options;
-      import_options.scope = "episode";
-      import_options.bundle = episode;
-      import_options.dry_run = false;
-      const auto receipt = episode_import_bundle_impl(import_options);
-      ok = ok && receipt.value("ok", false);
-      receipts.push_back(receipt);
-      if (!receipt.value("ok", false))
-        break;
-    }
-    return {{"schema", "kungfu.facts.library-import/v1"},
-            {"ok", ok},
-            {"dry_run", false},
-            {"source_mode", text_or(bundle, "mode")},
-            {"receipt_count", receipts.size()},
-            {"receipts", receipts},
-            {"preflight_receipts", preflight_receipts}};
-  }
-
-  [[nodiscard]] nlohmann::json assessment_contract(const storage_service_options &options) const {
-    (void)options;
-    return trust::assessment_contract_json();
-  }
-
-  [[nodiscard]] nlohmann::json assessment_request(const storage_service_options &options) const {
-    return trust::request_assessment(options.runtime_dir, object_or_empty(options.operation_options, "request"),
-                                     int64_or(options.operation_options, "system_time"));
-  }
-
-  [[nodiscard]] nlohmann::json assessment_execute(const storage_service_options &options) const {
-    return trust::execute_assessment(options.runtime_dir, text_or(options.operation_options, "assessment_key"),
-                                     text_or(options.operation_options, "executor_profile", "process"),
-                                     int64_or(options.operation_options, "system_time"));
-  }
-
-  [[nodiscard]] nlohmann::json assessment_status(const storage_service_options &options) const {
-    return trust::query_assessment(options.runtime_dir, text_or(options.operation_options, "assessment_key"));
-  }
-
-  [[nodiscard]] nlohmann::json assessment_list(const storage_service_options &options) const {
-    return trust::list_assessments(options.runtime_dir);
-  }
-
-  [[nodiscard]] nlohmann::json assessment_invalidate(const storage_service_options &options) const {
-    return trust::invalidate_assessment(options.runtime_dir, text_or(options.operation_options, "assessment_key"),
-                                        text_or(options.operation_options, "changed_root"),
-                                        text_or(options.operation_options, "reason"),
-                                        int64_or(options.operation_options, "system_time"));
-  }
-
-  [[nodiscard]] nlohmann::json trust_require(const storage_service_options &options) const {
-    return trust::require_trust(options.runtime_dir, text_or(options.operation_options, "assessment_key"),
-                                text_or(options.operation_options, "purpose"));
-  }
-
-  [[nodiscard]] nlohmann::json episode_begin(const storage_service_options &options) const {
-    return episode_record_body_json(
-        episode_store(options).begin(parse_episode_begin_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_heartbeat(const storage_service_options &options) const {
-    return episode_record_body_json(
-        episode_store(options).heartbeat(parse_episode_heartbeat_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_end(const storage_service_options &options) const {
-    return render_episode_close_write_result(episode_store(options).end(
-        parse_episode_close_options(options.operation_options, yy_enums::EpisodeStatus::Ended)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_abort(const storage_service_options &options) const {
-    return render_episode_close_write_result(episode_store(options).abort(
-        parse_episode_close_options(options.operation_options, yy_enums::EpisodeStatus::Aborted)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_attach_frame(const storage_service_options &options) const {
-    return episode_record_body_json(
-        episode_store(options).attach_frame(parse_episode_frame_attach_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_attach_ref(const storage_service_options &options) const {
-    return episode_record_body_json(
-        episode_store(options).attach_ref(parse_episode_ref_attach_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_list(const storage_service_options &options) const {
-    return episode_store(options).list(uint64_or(options.operation_options, "location_uid"), options.limit);
-  }
-
-  [[nodiscard]] nlohmann::json episode_inspect(const storage_service_options &options) const {
-    auto inspected = episode_ref_store(options).store.inspect(options.episode_id);
-    auto qualification_options = options;
-    qualification_options.scope = "episode";
-    const auto fsck = default_storage_service().fsck(parse_storage_fsck_request(qualification_options));
-    if (fsck.qualification.has_value())
-      inspected["qualification"] = episode_qualification_json(*fsck.qualification);
-    return inspected;
-  }
-
-  [[nodiscard]] nlohmann::json episode_recover(const storage_service_options &options) const {
-    return render_episode_recover_result(
-        episode_store(options).recover(parse_episode_recover_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json episode_projection_rebuild(const storage_service_options &options) const {
-    return episode_projection_rebuild_json(episode_manifest_projection(options.runtime_dir).rebuild_typed());
-  }
-
-  [[nodiscard]] nlohmann::json source_register(const storage_service_options &options) const {
-    return source_registry_record_json(
-        source_registry_store(options).register_source(parse_source_register_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json source_update_head(const storage_service_options &options) const {
-    return source_registry_record_json(
-        source_registry_store(options).update_head(parse_source_head_update_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json source_record_accepted_range(const storage_service_options &options) const {
-    return source_registry_record_json(
-        source_registry_store(options).record_accepted_range(parse_accepted_range_options(options.operation_options)));
-  }
-
-  [[nodiscard]] nlohmann::json source_list(const storage_service_options &options) const {
-    return source_registry_store(options).list();
-  }
-
-  [[nodiscard]] nlohmann::json source_inspect(const storage_service_options &options) const {
-    return source_registry_store(options).inspect(text_or(options.operation_options, "source_id", options.source_id));
-  }
-
-  [[nodiscard]] nlohmann::json source_registry_fsck(const storage_service_options &options) const {
-    const auto source_id = text_or(options.operation_options, "source_id", options.source_id);
-    const auto journal = source_registry_store(options).fsck_typed(source_id);
-    const auto render_issue = [](const yy_storage::source_registry_fsck_issue &issue) {
-      nlohmann::json row = {{"code", issue.code}};
-      if (issue.source_uid.has_value())
-        row["source_uid"] = *issue.source_uid;
-      if (issue.source_id.has_value())
-        row["source_id"] = *issue.source_id;
-      if (issue.count.has_value())
-        row["count"] = *issue.count;
-      return row;
-    };
-    nlohmann::json errors = nlohmann::json::array();
-    for (const auto &issue : journal.errors)
-      errors.push_back(render_issue(issue));
-    nlohmann::json warnings = nlohmann::json::array();
-    for (const auto &issue : journal.warnings)
-      warnings.push_back(render_issue(issue));
-    const auto projection = source_registry_projection(options.runtime_dir).verify_typed();
-    // ADR-0037: fsck verifies journal + projection. The journal is the
-    // authority; a rebuildable projection that has drifted from it is degraded,
-    // not failed.
-    const bool projection_degraded = projection.status == "degraded";
-    return {{"ok", journal.ok && !projection_degraded},
-            {"status", !journal.ok ? "failed" : (projection_degraded ? "degraded" : "ok")},
-            {"schema", journal.schema},
-            {"runtime_dir", journal.runtime_dir},
-            {"authority", journal.authority},
-            {"errors", std::move(errors)},
-            {"warnings", std::move(warnings)},
-            {"checked", {{"source_registry_records", journal.source_registry_records}, {"sources", journal.sources}}},
-            {"projection", projection_verification_json(projection)}};
-  }
-
-  [[nodiscard]] nlohmann::json source_registry_rebuild(const storage_service_options &options) const {
-    return projection_rebuild_json(source_registry_projection(options.runtime_dir).rebuild_typed());
-  }
-};
-
 class file_storage_service : public storage_service {
 public:
   [[nodiscard]] storage_status_result status(const storage_status_request &request) const override {
@@ -5531,11 +3330,6 @@ public:
 
 const file_storage_service &typed_storage_service_instance() {
   static const file_storage_service service;
-  return service;
-}
-
-const file_storage_json_edge_service &storage_json_edge_service_instance() {
-  static const file_storage_json_edge_service service;
   return service;
 }
 
@@ -5865,7 +3659,9 @@ nlohmann::json make_request(storage_operation operation, const storage_service_o
   };
 }
 
-} // namespace
+} // namespace detail
+
+using namespace detail;
 
 size_t storage_query_result::row_count() const {
   return std::visit([](const auto &typed_rows) { return typed_rows.size(); }, rows);
@@ -6920,121 +4716,8 @@ nlohmann::json make_storage_service_request(const std::string &operation, const 
 
 nlohmann::json run_storage_service_operation(const std::string &operation, const std::string &runtime_dir,
                                              const nlohmann::json &options) {
-  const auto parsed_operation = parse_storage_operation(operation);
-  const auto parsed_options = parse_storage_service_options(runtime_dir, options);
-  switch (parsed_operation) {
-  case storage_operation::Status:
-    return storage_json_edge_service_instance().status(parsed_options);
-  case storage_operation::Fsck:
-    return storage_json_edge_service_instance().fsck(parsed_options);
-  case storage_operation::RepairPlan:
-    return storage_json_edge_service_instance().repair_plan(parsed_options);
-  case storage_operation::RepairFetch:
-    return storage_json_edge_service_instance().repair_fetch(parsed_options);
-  case storage_operation::RepairApply:
-    return storage_json_edge_service_instance().repair_apply(parsed_options);
-  case storage_operation::ExportBundle:
-    return storage_json_edge_service_instance().export_bundle(parsed_options);
-  case storage_operation::ImportBundle:
-    return storage_json_edge_service_instance().import_bundle(parsed_options);
-  case storage_operation::RebuildIndex:
-    return storage_json_edge_service_instance().rebuild_index(parsed_options);
-  case storage_operation::GcPlan:
-    return storage_json_edge_service_instance().gc_plan(parsed_options);
-  case storage_operation::CompactPlan:
-    return storage_json_edge_service_instance().compact_plan(parsed_options);
-  case storage_operation::VerifySync:
-    return storage_json_edge_service_instance().verify_sync(parsed_options);
-  case storage_operation::Query:
-    return render_storage_query_result(default_storage_service().query(parse_storage_query_request(parsed_options)));
-  case storage_operation::QueryPlan:
-    return storage_json_edge_service_instance().query_plan(parsed_options);
-  case storage_operation::FactQuery:
-    return storage_json_edge_service_instance().fact_query(parsed_options);
-  case storage_operation::FactChangelog:
-    return storage_json_edge_service_instance().fact_changelog(parsed_options);
-  case storage_operation::SavedQueryCatalog:
-    return storage_json_edge_service_instance().saved_query_catalog(parsed_options);
-  case storage_operation::ProfileLifecycle:
-    return storage_json_edge_service_instance().profile_lifecycle(parsed_options);
-  case storage_operation::KfxRuntime:
-    return storage_json_edge_service_instance().kfx_runtime(parsed_options);
-  case storage_operation::FactContract:
-    return storage_json_edge_service_instance().fact_contract(parsed_options);
-  case storage_operation::FactDeclareWorld:
-    return storage_json_edge_service_instance().fact_declare_world(parsed_options);
-  case storage_operation::FactDeclareSurface:
-    return storage_json_edge_service_instance().fact_declare_surface(parsed_options);
-  case storage_operation::FactObserve:
-    return storage_json_edge_service_instance().fact_observe(parsed_options);
-  case storage_operation::FactState:
-    return storage_json_edge_service_instance().fact_state(parsed_options);
-  case storage_operation::FactLibraryContract:
-    return storage_json_edge_service_instance().fact_library_contract(parsed_options);
-  case storage_operation::FactTypeCreate:
-    return storage_json_edge_service_instance().fact_type_create(parsed_options);
-  case storage_operation::FactTypeList:
-    return storage_json_edge_service_instance().fact_type_list(parsed_options);
-  case storage_operation::FactMaterialPut:
-    return storage_json_edge_service_instance().fact_material_put(parsed_options);
-  case storage_operation::FactMaterialList:
-    return storage_json_edge_service_instance().fact_material_list(parsed_options);
-  case storage_operation::FactLibraryExport:
-    return storage_json_edge_service_instance().fact_library_export(parsed_options);
-  case storage_operation::FactLibraryImport:
-    return storage_json_edge_service_instance().fact_library_import(parsed_options);
-  case storage_operation::AssessmentContract:
-    return storage_json_edge_service_instance().assessment_contract(parsed_options);
-  case storage_operation::AssessmentRequest:
-    return storage_json_edge_service_instance().assessment_request(parsed_options);
-  case storage_operation::AssessmentExecute:
-    return storage_json_edge_service_instance().assessment_execute(parsed_options);
-  case storage_operation::AssessmentStatus:
-    return storage_json_edge_service_instance().assessment_status(parsed_options);
-  case storage_operation::AssessmentList:
-    return storage_json_edge_service_instance().assessment_list(parsed_options);
-  case storage_operation::AssessmentInvalidate:
-    return storage_json_edge_service_instance().assessment_invalidate(parsed_options);
-  case storage_operation::TrustRequire:
-    return storage_json_edge_service_instance().trust_require(parsed_options);
-  case storage_operation::Layout:
-    return storage_json_edge_service_instance().layout(parsed_options);
-  case storage_operation::EpisodeBegin:
-    return storage_json_edge_service_instance().episode_begin(parsed_options);
-  case storage_operation::EpisodeHeartbeat:
-    return storage_json_edge_service_instance().episode_heartbeat(parsed_options);
-  case storage_operation::EpisodeEnd:
-    return storage_json_edge_service_instance().episode_end(parsed_options);
-  case storage_operation::EpisodeAbort:
-    return storage_json_edge_service_instance().episode_abort(parsed_options);
-  case storage_operation::EpisodeAttachFrame:
-    return storage_json_edge_service_instance().episode_attach_frame(parsed_options);
-  case storage_operation::EpisodeAttachRef:
-    return storage_json_edge_service_instance().episode_attach_ref(parsed_options);
-  case storage_operation::EpisodeList:
-    return storage_json_edge_service_instance().episode_list(parsed_options);
-  case storage_operation::EpisodeInspect:
-    return storage_json_edge_service_instance().episode_inspect(parsed_options);
-  case storage_operation::EpisodeRecover:
-    return storage_json_edge_service_instance().episode_recover(parsed_options);
-  case storage_operation::EpisodeProjectionRebuild:
-    return storage_json_edge_service_instance().episode_projection_rebuild(parsed_options);
-  case storage_operation::SourceRegister:
-    return storage_json_edge_service_instance().source_register(parsed_options);
-  case storage_operation::SourceUpdateHead:
-    return storage_json_edge_service_instance().source_update_head(parsed_options);
-  case storage_operation::SourceRecordAcceptedRange:
-    return storage_json_edge_service_instance().source_record_accepted_range(parsed_options);
-  case storage_operation::SourceList:
-    return storage_json_edge_service_instance().source_list(parsed_options);
-  case storage_operation::SourceInspect:
-    return storage_json_edge_service_instance().source_inspect(parsed_options);
-  case storage_operation::SourceRegistryFsck:
-    return storage_json_edge_service_instance().source_registry_fsck(parsed_options);
-  case storage_operation::SourceRegistryRebuild:
-    return storage_json_edge_service_instance().source_registry_rebuild(parsed_options);
-  }
-  throw std::invalid_argument("unknown storage operation");
+  return detail::dispatch_json_edge_operation(parse_storage_operation(operation),
+                                              parse_storage_service_options(runtime_dir, options));
 }
 
 nlohmann::json accept_storage_manifest(const std::string &runtime_dir, const nlohmann::json &manifest) {
@@ -7091,11 +4774,10 @@ std::string write_storage_payload_bytes(const std::string &runtime_dir, const st
   }
   const auto provider = shared_provider(runtime_dir);
   provider->write_payload(digest, raw);
-  return provider->name() == PROVIDER_ROCKSDB ? storage_uri(PROVIDER_ROCKSDB, runtime_dir, "payloads/" + digest)
-                                              : payload_path(runtime_dir, digest).string();
+  return payload_uri_for(provider->name(), runtime_dir, digest);
 }
 
-namespace {
+namespace detail {
 
 nlohmann::json content_result_json(const yy_storage::content_store_result &result) {
   return {{"ok", result.ok()},
@@ -7105,6 +4787,10 @@ nlohmann::json content_result_json(const yy_storage::content_store_result &resul
           {"existed", result.existed},
           {"message", result.message}};
 }
+
+} // namespace detail
+
+namespace {
 
 // Accept "<algo>:<hex>" or bare hex, mirroring the kernel's ref resolution.
 bool parse_content_hash_text(const std::string &text, yy_storage::content_hash &hash, std::string &message) {
@@ -7209,13 +4895,13 @@ nlohmann::json storage_service_capabilities() {
                         {{"name", PROVIDER_FILE},
                          {"default", provider.name == PROVIDER_FILE},
                          {"selected", provider.name == PROVIDER_FILE},
-                         {"layout", provider_layout_json(file_storage_provider("").layout())},
-                         {"runtime", provider_runtime_json(file_storage_provider("").runtime())}},
+                         {"layout", provider_layout_json(provider_layout_for(PROVIDER_FILE))},
+                         {"runtime", provider_runtime_json(provider_runtime_for(PROVIDER_FILE))}},
                         {{"name", PROVIDER_ROCKSDB},
                          {"default", provider.name == PROVIDER_ROCKSDB},
                          {"selected", provider.name == PROVIDER_ROCKSDB},
-                         {"layout", provider_layout_json(rocksdb_storage_provider("").layout())},
-                         {"runtime", provider_runtime_json(rocksdb_storage_provider("").runtime())}},
+                         {"layout", provider_layout_json(provider_layout_for(PROVIDER_ROCKSDB))},
+                         {"runtime", provider_runtime_json(provider_runtime_for(PROVIDER_ROCKSDB))}},
                     })},
       {"projections",
        nlohmann::json::array(
